% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{aaa_Get_Started}
\alias{aaa_Get_Started}
\alias{get_started}
\title{Get URL of Metabodecon "Get Started" Page}
\usage{
aaa_Get_Started(open_browser = interactive())

get_started(open_browser = interactive())
}
\arguments{
\item{open_browser}{If TRUE, the "Get Stated" page is opened in the default
browser.}
}
\value{
A character string containing the URL of the "Get Started" page.
}
\description{
\code{get_started} and \code{aaa_Get_Started} both return (and optionally open) the URL
of the "Get Started" page of the metabodecon documentation. The
\code{aaa_Get_Started} version exists, because functions are listed alphabetically
in the reference manual and we want \code{get_started} to be shown at the top of
the list (i.e., it needs to start with an 'a').
}
\examples{
get_started(open_browser = FALSE)
get_started()
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
