% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_generate_lorentz_curves_v12.R
\name{add_return_list_v12}
\alias{add_return_list_v12}
\title{create backwards compatible return list}
\usage{
add_return_list_v12(
  spec = glc_urine1_yy_ni3_dbg()$rv$urine_1,
  n = 1,
  nam = "urine_1",
  debug = TRUE
)
}
\arguments{
\item{spec}{Deconvoluted spectrum as returned by \code{\link[=refine_lorentz_curves_v12]{refine_lorentz_curves_v12()}}.}

\item{n}{Number of deconvoluted spectrum.}

\item{nam}{Name of current spectrum.}

\item{debug}{Add debug info to the return list}
}
\description{
create backwards compatible return list
}
