% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{AlignedSpectra}
\alias{AlignedSpectra}
\title{AlignedSpectra Class}
\usage{
AlignedSpectra(...)
}
\description{
Each object of this class:
\itemize{
\item Represents a set of one or more one-dimensional, deconvoluted and aligned NMR spectra.
\item Is a \link{data.frame} where element i,j gives the integral value of the signal from spectrum i and chemical shift j.
Objects of this class can be created using \code{\link[=align_spectra]{align_spectra()}}.
}
}
