% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R, R/rust.R
\name{as_rust_spectrum}
\alias{as_rust_spectrum}
\title{Convert to Rust Spectrum}
\usage{
as_rust_spectrum(spectrum, sfr)

as_rust_spectrum(spectrum, sfr)
}
\arguments{
\item{spectrum}{The spectrum to be converted.}

\item{sfr}{The spectral frequency range.}
}
\value{
A Rust spectrum object.

A Rust spectrum object.
}
\description{
Converts a given spectrum to a Rust spectrum.

Converts a given spectrum to a Rust spectrum.
}
