% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{as_spectra}
\alias{as_spectra}
\alias{as_decon2}
\alias{as_decons2}
\title{Convert to a Metabodecon Object}
\usage{
as_spectra(
  x,
  file_format = "bruker",
  expno = 10,
  procno = 10,
  raw = FALSE,
  silent = TRUE,
  force = FALSE
)

as_decon2(x)

as_decons2(x)
}
\arguments{
\item{x}{The object to convert. For \code{\link[=as_spectra]{as_spectra()}}, either a \code{spectrum},
a list of \code{spectrum}, or a path passed to \code{\link[=read_spectra]{read_spectra()}}.}

\item{file_format, expno, procno, raw, silent, force}{Passed to \code{\link[=read_spectra]{read_spectra()}} when \code{x} is a path.}
}
\value{
An object of the requested class.
}
\description{
Identity-or-validate converters between metabodecon classes. See
\link{metabodecon-classes} for the class hierarchy.
}
\examples{
as_spectra(sim[[1]])
as_decon2(deconvolute(sim[[1]], sfr = c(3.55, 3.35)))
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
