% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cache_example_datasets_bak}
\alias{cache_example_datasets_bak}
\title{Cache Example Files}
\usage{
cache_example_datasets_bak(persistent = NULL, ask = interactive())
}
\arguments{
\item{persistent}{Logical. If TRUE, the datasets are cached permanently. If NULL, the function prompts the user. Default is NULL.}

\item{ask}{Logical. If TRUE, the function operates in interactive mode and prompts the user for input. Default is interactive().}
}
\value{
The path to the cached datasets.
}
\description{
This function caches example datasets to speed up future calls. It checks if the file exists and has the correct size. If not, it prompts the user for action.
}
\examples{
cache_example_datasets(persistent = FALSE)
}
