% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{coef.mdlm}
\alias{coef.mdlm}
\title{Extract mdlm coefficients}
\usage{
\method{coef}{mdlm}(object, ...)
}
\arguments{
\item{object}{Fitted \code{mdlm} object.}

\item{...}{Additional arguments passed to \code{stats::coef}.}
}
\value{
Coefficient object for lasso.
}
\description{
Returns lasso coefficients for \code{mdlm} objects.
}
\examples{
\dontrun{
  m <- cv_mdlm(spectra, y, sfr = c(11, -2))
  coef(m)
}
}
