% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{coef.mdm}
\alias{coef.mdm}
\title{Extract mdm coefficients}
\usage{
\method{coef}{mdm}(object, ...)
}
\arguments{
\item{object}{Fitted \code{mdm} object.}

\item{...}{Additional arguments passed to \code{stats::coef}.}
}
\value{
Coefficient object for lasso, otherwise \code{NULL}.
}
\description{
Returns lasso coefficients for \code{mdm} objects fitted with lasso.
}
\examples{
\dontrun{
  m <- cv_mdm(spectra, y, sfr = c(11, -2))
  coef(m)
}
}
