% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_v2.R
\name{collapse}
\alias{collapse}
\title{Collapse a vector into a string}
\usage{
collapse(x, sep = ", ")
}
\arguments{
\item{x}{A vector to collapse.}

\item{sep}{A string to use as the separator between elements. Default is ", ".}
}
\value{
A single string with elements of x separated by sep.
}
\description{
Collapses a vector into a single string, with elements separated by a specified separator. Essentially a shorthand for \code{paste(x, collapse = sep)}.
}
\examples{
\dontrun{
collapse(c("a", "b", "c")) # "a, b, c"
collapse(1:5, sep = "-")   # "1-2-3-4-5"
}
}
\keyword{internal}
