% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_as.R
\name{convert}
\alias{convert}
\alias{as_spectrum}
\alias{as_gspec}
\alias{as_gdecon}
\alias{as_decon1}
\alias{as_spectra}
\title{Convert to a Metabodecon Class}
\usage{
convert(x, to)

as_spectrum(x, sf = c(1000, 1e+06))

as_gspec(x, sf = c(1000, 1e+06))

as_gdecon(x)

as_decon1(x, sf = NULL)

as_spectra(
  x,
  ...,
  nams = NULL,
  file_format = "bruker",
  expno = 10,
  procno = 10,
  raw = FALSE,
  silent = TRUE,
  force = FALSE
)
}
\arguments{
\item{x}{The object to convert.}

\item{to}{The class to convert to. For details see \link{metabodecon_classes}.}

\item{nams}{Names to overwrite the individual spectrum names with. Must be NULL or a character vector of same length as \code{list(x, ...)}. If not provided, the names of the individual spectrum objects will be used, if available, else a default name of the form "spectrum_\%d" where "\%d" equals the spectrum number.#' @return An object of the specified class.}
}
\description{
Convert an object to a Metabodecon class.
}
\examples{
dirpath <- metabodecon_file("sim_subset")
spectra <- read_spectra(dirpath)
gspecs <- as_gspecs(spectra)
gdecons <- deconvolute_gspecs(gspecs)
gspec <- as_gspec(spectrum)
gdecon <- deconvolute_gspec(gspec)
decon2 <- as_decon2(gdecon)
}
