% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_util.R
\name{convert_pos}
\alias{convert_pos}
\title{Convert positions from unit A to unit B}
\usage{
convert_pos(xa, wa, wb, x0a, x0b)
}
\arguments{
\item{xa}{A numeric vector with positions in unit A.}

\item{wa}{Width of a certain interval (e.g. the spectrum width) in unit A.}

\item{wb}{With of the same interval in unit B.}

\item{x0a}{The position of a reference point in unit A (e.g. the most right point of the spectrum).}

\item{x0b}{The position of the same reference point in unit B.}
}
\value{
A numeric vector of positions converted from unit A to unit B.
}
\description{
Converts positions from unit A to unit B using a conversion factor.
}
\examples{
\donttest{
urine_1 <- pkg_file("example_datasets/bruker/urine/urine_1")
deconv <- generate_lorentz_curves(urine_1, ask = FALSE)[[1]]
sdp <- deconv$x_values
ppm <- deconv$x_values_ppm
peak9_sdp <- deconv$x_0[1]
peak9_ppm <- convert_pos(peak9_sdp,
    width(sdp), width(ppm),  # width of the spectrum in both units
    sdp[3], ppm[3]           # position of any reference point in both units
)
stopifnot(all.equal(peak9_ppm, deconv$x_0_ppm[9]))
}
}
