% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R, R/rust.R
\name{convert_to_decon2}
\alias{convert_to_decon2}
\title{Convert to Decon2}
\usage{
convert_to_decon2(spectrum, deconvoluter, deconvolution, sfr)

convert_to_decon2(spectrum, deconvoluter, deconvolution, sfr)
}
\arguments{
\item{spectrum}{The original spectrum.}

\item{deconvoluter}{The deconvoluter object.}

\item{deconvolution}{The deconvolution results.}

\item{sfr}{The spectral frequency range.}
}
\value{
A Decon2 object.

A Decon2 object.
}
\description{
Converts a given spectrum and deconvolution results to Decon2 format.

Converts a given spectrum and deconvolution results to Decon2 format.
}
