% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_util.R
\name{convert_width}
\alias{convert_width}
\alias{convert_pos}
\title{Convert from unit A to unit B}
\usage{
convert_width(xa, ya, yb)

convert_pos(xa, ya, yb)
}
\arguments{
\item{xa}{For \code{convert_width} a numeric vector specifying widths in unit A. For \code{convert_pos} a numeric vector specifying positions in unit A.}

\item{ya}{A numeric vector giving the positions of at least two points in unit A.}

\item{yb}{A numeric vector giving the positions of the same points in unit B.}
}
\value{
A numeric vector of values converted from unit A to unit B.
}
\description{
\code{convert_pos} converts positions from unit A to unit B. \code{convert_width} converts widths from unit A to unit B. If the direction of units A and B is reversed, the width's sign will be reversed as well. To keep widths strictly positive, wrap the result with \code{abs()}.
}
\examples{
ya <- c(244, 246, 248, 250, 252)
yb <- c(15, 10, 5, 0, -5)
convert_width(c(2, 4, 8), ya, yb)
convert_pos(c(247, 249), ya, yb)
}
