% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_test_helpers.R
\name{count_stretches}
\alias{count_stretches}
\title{Count Stretches of Increases and Decreases}
\usage{
count_stretches(x)
}
\arguments{
\item{vec}{A numeric vector.}
}
\value{
A numeric vector containing the lengths of stretches of increases and decreases.
}
\description{
Counts the lengths of consecutive increases and decreases in a numeric vector.
}
\examples{
#
# Example Data (x)
# | 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8 |
# |-------------------------------|
# |   |   |   |###|   |   |   |###|
# |   |   |###|###|   |   |###|###|
# |   |###|###|###|###|   |###|###|
# |###|###|###|###|###|###|###|###|
# |-------------------------------|
# |   +   +   +   -   -   +   +   |
# |-------------------------------|
#
x <- c(1.0, 2.2, 3.0, 4.4, 2.0, 1.0, 3.0, 4.0)
count_stretches(x) # Returns c(3, 2, 2) because we have 3+, 2-, 2+
}
