% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir.R
\name{datadir}
\alias{datadir}
\title{Return path to metabodecon's data directory}
\usage{
datadir(file = "", warn = TRUE, persistent = NULL)
}
\arguments{
\item{file}{Relative path to a file within the data directory.}

\item{warn}{Print a warning message when the requested path does not yet exist?}

\item{persistent}{Return the path to the persistent data directory instead of the temporary one?}
}
\description{
Returns the path to the directory where \code{\link[=download_example_datasets]{download_example_datasets()}} stores metabodecon's example data sets or any file within that directory.
By default this directory is a subdirectory of R's temporary session directory.
If \code{persistent} is set to \code{TRUE}, the directory equals the data directory returned by \code{\link[tools:userdir]{tools::R_user_dir()}} instead.
}
\details{
The decision to use a temporary data dir as default and a persistent one only optionally was made to conform to CRAN package policies, which state that:
\emph{Packages should not write in the user's home filespace (including clipboards), nor anywhere else on the file system apart from the R session's temporary directory (or during installation in the location pointed to by TMPDIR: and such usage should be cleaned up). Installing into the system's R installation (e.g., scripts to its bin directory) is not allowed.}

\emph{Limited exceptions may be allowed in interactive sessions if the package obtains confirmation from the user.}

\emph{For R version 4.0 or later (hence a version dependency is required or only conditional use is possible), packages may store user-specific data, configuration and cache files in their respective user directories obtained from tools::R_user_dir(), provided that by default sizes are kept as small as possible and the contents are actively managed (including removing outdated material).}

Source: \href{https://cran.r-project.org/web/packages/policies.html}{cran.r-project.org/web/packages/policies}
}
\examples{
# Return path to persistent data dir if it exists, else path to temp data dir
datadir()

# Return path to temporary data directory
datadir(persistent = FALSE)

# Return path to persistent data directory
datadir(persistent = TRUE)

# Return path to "<persistent-data-dir>/bruker/urine" if it exists.
# Else return path to "<temp-data-dir>/bruker/urine"
datadir(file = "bruker/urine")
}
\seealso{
\code{\link[=download_example_datasets]{download_example_datasets()}}

\code{\link[=datadir_persistent]{datadir_persistent()}}, \code{\link[=datadir_temp]{datadir_temp()}}
}
