% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_datadir.R
\name{datadir_persistent}
\alias{datadir_persistent}
\title{Persistent Data Directory}
\usage{
datadir_persistent()
}
\value{
Path to the persistent data directory.
}
\description{
Returns the path to the persistent data directory where metabodecon's data sets are stored. This directory equals the data directory returned by \code{\link[tools:userdir]{tools::R_user_dir()}} plus additional path normalization.
}
\seealso{
\code{\link[=datadir]{datadir()}}, \code{\link[=datadir_temp]{datadir_temp()}}
}
