% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/58_data.R
\name{datadir_temp}
\alias{datadir_temp}
\title{Temporary Data Directory}
\usage{
datadir_temp()
}
\value{
Returns the path to the temporary data directory.
}
\description{
Returns the path to the temporary data directory where
metabodecon's data sets are stored. This directory equals subdirectory 'data'
of metabodecons temporary session directory \code{\link[=tmpdir]{tmpdir()}} plus additional path
normalization.
}
\examples{

datadir_temp()
}
\seealso{
\code{\link[=tmpdir]{tmpdir()}}, \code{\link[=datadir]{datadir()}}, \code{\link[=datadir_persistent]{datadir_persistent()}}
}
