% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R, R/rust.R
\name{deconvolute_spectra_rust}
\alias{deconvolute_spectra_rust}
\title{Multi Deconvolute Spectra using Rust Backend}
\usage{
deconvolute_spectra_rust(
  spectra,
  sfr,
  nfit = 3,
  smopts = c(2, 5),
  delta = 6.4,
  ignore_regions = NULL,
  parallel = FALSE,
  optimize_settings = FALSE
)

deconvolute_spectra_rust(
  spectra,
  sfr,
  nfit = 3,
  smopts = c(2, 5),
  delta = 6.4,
  ignore_regions = NULL,
  parallel = FALSE,
  optimize_settings = FALSE
)
}
\arguments{
\item{spectra}{A list of spectra to be deconvoluted.}

\item{sfr}{The spectral frequency range.}

\item{nfit}{Number of iterations for the analytical fitter. Default is 3.}

\item{smopts}{Smoothing options as a vector of two elements: iterations and window size. Default is c(2, 5).}

\item{delta}{Noise score threshold. Default is 6.4.}

\item{ignore_regions}{Regions to ignore during deconvolution. Default is NULL.}

\item{parallel}{Whether to perform deconvolution in parallel. Default is FALSE.}

\item{optimize_settings}{Whether to optimize settings. Default is FALSE.}
}
\value{
A list of deconvoluted spectra and additional information.

A list of deconvoluted spectra and additional information.
}
\description{
Deconvolutes a list of spectra using the Rust backend.

Deconvolutes a list of spectra using the Rust backend.
}
