% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{DeconvolutedSpectra}
\alias{DeconvolutedSpectra}
\alias{is_decon_list}
\title{DeconvolutedSpectra Class}
\usage{
is_decon_list(x)
}
\description{
Each object of this class:
\itemize{
\item Represents one or more one-dimensional, deconvoluted NMR spectra.
\item Is a list of \link{DeconvolutedSpectrum} objects.
Objects of this class can be created using \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}.
}
}
\examples{
TODO
}
