% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_deconvolute.R
\name{determine_water_signal_v20}
\alias{determine_water_signal_v20}
\title{Calculate water signal parameters}
\usage{
determine_water_signal_v20(spec, hwidth_ppm, bwc = TRUE, ask = TRUE)
}
\arguments{
\item{spec}{A list representing the spectrum.}

\item{hwidth_ppm}{The half width in ppm. Default is wshw.}

\item{bwc}{Use the old, slightly incorrect methods for calculating water signal values to maintain backwards compatibility with MetaboDecon1D results? For details see issue \verb{Check: water signal calculation} in \code{TODOS.md}.}
}
\value{
List of parameters including half width in dp and ppm, center line in dp and ppm and right and left borders in dp and ppm.
}
\description{
Calculates water signal parameters for a given spectrum.
}
