% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_alignment.R
\name{dohCluster}
\alias{dohCluster}
\title{dohCluster}
\usage{
dohCluster(
  X,
  peakList,
  refInd = 0,
  maxShift = 100,
  acceptLostPeak = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{X}{\code{data_matrix} as returned by \code{\link[=gen_feat_mat]{gen_feat_mat()}}}

\item{peakList}{\code{peakList} as returned by function \code{\link[=gen_feat_mat]{gen_feat_mat()}}}

\item{refInd}{Number of the reference spectrum i.e. the spectrum to which all signals will be aligned to.}

\item{maxShift}{Maximum number of points along the "ppm-axis" which a value can be moved (e.g. 50).}

\item{acceptLostPeak}{(logic) default is TRUE}

\item{verbose}{(logic) default is TRUE}
}
\value{
A list containing two data frames \code{Y} and \code{new_peakList}. The first one contains the aligned spectra, the second one contains the aligned signals of each spectrum.
}
\description{
Helper of \code{\link[=speaq_align]{speaq_align()}}.
}
\details{
The function dohCluster of the speaq package has been rewritten to be compatible with the data generated by MetaboDecon1D and the function gen_feat_mat and to return a new peakList of aligned spectra. Overwrite original dohCluster function of speaq package. Function is able to return the new peakList of aligned spectra.
}
\examples{
\dontrun{
res <- dohCluster(X, peakList, refInd = 0, maxShift = 100, acceptLostPeak = TRUE, verbose = TRUE)
}
}
