% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_example_datasets}
\alias{download_example_datasets}
\title{Download metabodecon example datasets}
\usage{
download_example_datasets(
  dst_dir = NULL,
  extract = TRUE,
  cache_persistent = NULL,
  ask = interactive()
)
}
\description{
Downloads example datasets for testing metabodecon's functionality. Due to the size constraints for R packages, these datasets are not included by default when the package is installed.
Datasets are download and extracted to the directory returned by \code{\link[=datadir]{datadir()}}.
By default this directory is a subdirectory of R's temporary session directory.
If \code{persistent} is set to \code{FALSE}, the directory equals the one returned by \code{\link[=datadir_temp]{datadir_temp()}}
If \code{persistent} is set to \code{TRUE}, the directory equals the one returned by \code{\link[=datadir_persistent]{datadir_persistent()}}.
If \code{persistent} is set to \code{NULL} and \code{ask} is TRUE, the user is asked which one he wants to use.
}
\examples{
\dontrun{
download_example_datasets()
}
}
\seealso{
\code{\link[=datadir]{datadir()}}
}
