% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_example_datasets_zip}
\alias{download_example_datasets_zip}
\title{Download Example Datasets Zip}
\usage{
download_example_datasets_zip(path, copyfrom = NULL)
}
\arguments{
\item{path}{A string. The path where the downloaded file will be saved.}
}
\value{
The path where the downloaded file is saved.
}
\description{
This function downloads the example_datasets.zip file from the specified URL and saves it to the destination directory provided. If the directory does not exist, it will be created. If the file already exists in the directory, it will be overwritten without asking for permission.
}
\examples{
\dontrun{
download_example_datasets_zip("/path/to/your/directory/example_datasets.zip")
}
}
