% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_generate_lorentz_curves.R
\name{generate_lorentz_curves}
\alias{generate_lorentz_curves}
\title{Generate Lorentz Curves from NMR Spectra}
\usage{
generate_lorentz_curves(
  data_path = file.path(download_example_datasets(), "bruker/urine"),
  file_format = "bruker",
  make_rds = FALSE,
  expno = 10,
  procno = 10,
  nfit = 10,
  wshw = 0.1527692,
  sfr = c(11.44494, -1.8828),
  smopts = c(2, 5),
  delta = 6.4,
  sf = c(1000, 1e+06),
  ask = TRUE,
  debug = FALSE,
  ncores = "auto"
)
}
\arguments{
\item{data_path}{Either the path to an existing directory containing measured NMR spectra or a dataframe with columns \code{ppm} (parts per million) and \code{si} (signal intensity) or a list of such dataframes.}

\item{file_format}{Format of the spectra files. Either \code{"bruker"} or \code{"jcampdx"}. Only relevant if \code{data_path} is a directory.}

\item{make_rds}{Store results as rds file on disk? Should be set to TRUE if many spectra are evaluated to decrease computation time.}

\item{expno}{The experiment number for the spectra files. E.g. \code{"10"}. Only relevant if \code{data_path} is a directory and \code{file_format} is \code{"bruker"}.}

\item{procno}{The processing number for the spectra. E.g. \code{"10"}. Only relevant if \code{data_path} is a directory and \code{file_format} is \code{"bruker"}.}

\item{nfit}{Number of iterations for the approximation of the parameters for the Lorentz curves.}

\item{wshw}{Half width of the water artefact in ppm.}

\item{sfr}{Row vector with two entries consisting of the ppm positions for the left and right border of the signal free region of the spectrum.}

\item{smopts}{Vector with two entries consisting of the number of smoothing iterations and the number of data points to use for smoothing (must be uneven).}

\item{delta}{Threshold value to distinguish between signal and noise.}

\item{sf}{Vector with two entries consisting of the factor to scale the x-axis and the factor to scale the y-axis.}

\item{ask}{Whether to ask for user input during the deconvolution process. If set to FALSE, the provided default values will be used.}

\item{debug}{Whether to return additional intermediate results for debugging purposes.}

\item{ncores}{Number of cores to use for parallel processing. If set to \code{"auto"}, the number of cores will be determined automatically. If set to a number greater than 1, the number of cores will be limited to the number of spectra or 1 if the operating system is Windows.}
}
\description{
Deconvolutes NMR spetra and generates a Lorentz curve for each detected signal within a spectra.
}
\details{
First, an automated curvature based signal selection is performed. Each signal is represented by 3 data points to allow the determination of initial Lorentz curves. These Lorentz curves are then iteratively adjusted to optimally approximate the measured spectrum.
}
\examples{
\donttest{
xp <- download_example_datasets()
dp <- file.path(xp, "bruker/urine")
ff <- "bruker"
x <- generate_lorentz_curves(dp, ff, ask = FALSE, nfit = 3, ncores = 1)
x2 <- generate_lorentz_curves(dp, ff, ask = FALSE, nfit = 3)
}
}
