% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/30_generate_lorentz_curves_v10.R
\name{generate_lorentz_curves_v10}
\alias{generate_lorentz_curves_v10}
\title{Generate Lorentz Curves}
\usage{
generate_lorentz_curves_v10(
  data_path,
  file_format = "bruker",
  make_rds = FALSE
)
}
\arguments{
\item{data_path}{Path to the parent folder where the original spectra are stored. After deconvolution this folder also contains for each spectrum two .txt files which contain for each spectrum the spectrum approximated from all deconvoluted signals and a parameter file that contains all numerical values of the deconvolution.}

\item{file_format}{Either \code{"bruker"} or \code{"jcampdx"}.}

\item{make_rds}{Store results as file \code{"spectrum_data.rds"} in folder \code{data_path}? If many spectra are evaluated and therefore computing time increases, this should be set to TRUE.}
}
\description{
Starts the actual deconvolution of your spectra and fits a Lorentz curve for each detected signal.
}
\details{
First, an automated curvature based signal selection is performed. Each signal is represented by 3 data points to allow the determination of initial Lorentz curves. These Lorentz curves are then iteratively adjusted to optimally approximate the measured spectrum.
}
\examples{
\donttest{
xds_path <- download_example_datasets()
data_path <- file.path(xds_path, "bruker/urine")
spectrum_data <- generate_lorentz_curves(data_path = data_path, file_format = "bruker")
}
}
