% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Retrieve directory path of an example dataset}
\usage{
get_data_dir(dataset_name = c("", "test", "urine"), warn = TRUE)
}
\arguments{
\item{dataset_name}{Either an string, "test" or "urine".}

\item{warn}{Whether to print a warning message when the example folders do
not yet exist, i.e. \code{\link[=download_example_datasets]{download_example_datasets()}} has not been called yet.}
}
\description{
This function return the path to the directory storing the
example files shipped with metabodecon.
}
\examples{
get_data_dir("urine")
}
\seealso{
\code{\link[=download_example_datasets]{download_example_datasets()}}
}
