% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get_names}
\alias{get_names}
\title{Returns the names of a metabodecon collection object.}
\usage{
get_names(x, default = "spectrum_\%d")
}
\arguments{
\item{x}{A metabodecon collection object.}

\item{default}{Default names if no names are found. Passed on to \code{get_default_names}.}
}
\value{
A character vector of names.
}
\description{
Returns the names of a metabodecon collection object.
}
\examples{
s1 <- list()
s2 <- list(name = "foo")
s3 <- list(name = "foo", meta = list(name = "bar"))

get_names(list(s1, s1)) # c("spectrum_1", "spectrum_2")
get_names(list(s1, myspec = s1)) # c("spectrum_1", "myspec")
get_names(list(s1, myspec = s2)) # c("spectrum_1", "foo")
get_names(list(s1, myspec = s3)) # c("spectrum_1", "bar")
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
