% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_v2.R
\name{get_num_input}
\alias{get_num_input}
\title{Get numeric input from user}
\usage{
get_num_input(prompt, min = -Inf, max = Inf, int = FALSE)
}
\arguments{
\item{prompt}{The prompt to display to the user.}

\item{min}{The minimum valid value. Default is -Inf.}

\item{max}{The maximum valid value. Default is Inf.}

\item{int}{Whether the input should be an integer. Default is FALSE.}
}
\value{
The user's input as a numeric value.
}
\description{
Prompts the user for input and checks if the input is a number between a minimum and maximum value. If the input is not valid, it keeps asking the user for input until they provide a valid response.
}
\examples{
\dontrun{
get_num_input("Enter a number between 1 and 10: ", min = 1, max = 10)
}
}
\keyword{internal}
