% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_alignment.R
\name{get_ppm_range}
\alias{get_ppm_range}
\title{Get ppm range}
\usage{
get_ppm_range(spectrum_data, show = FALSE)
}
\arguments{
\item{spectrum_data}{A list of deconvoluted spectra as returned by \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}.}

\item{show}{Whether to plot the ppm range on the spectrum plot.}
}
\value{
A vector containing the lowest and highest ppm value over all peaks of the provided deconvoluted spectra.
}
\description{
Returns the ppm range across all peaks of the provided deconvoluted spectra.
}
\examples{
sim <- metabodecon_file("bruker/sim_subset")
decons <- generate_lorentz_curves(
     sim, sfr = c(3.58, 3.42), wshw = 0,
     ask = FALSE, verbose = FALSE
)
ppm_rng <- get_ppm_range(decons, show = TRUE)
print(ppm_rng)
}
\author{
Wolfram Gronwald, Tobias Schmidt
}
