% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_v2.R
\name{get_str_input}
\alias{get_str_input}
\alias{get_yn_input}
\title{Get string input from user}
\usage{
get_str_input(prompt, valid)

get_yn_input(prompt)
}
\arguments{
\item{prompt}{The prompt to display to the user.}

\item{valid}{A vector of valid responses.}
}
\value{
The user's input.
}
\description{
Prompts the user for input and checks if the input is in a list of valid responses. If the input is not valid, it keeps asking the user for input until they provide a valid response.
}
\examples{
\dontrun{
get_str_input("Enter a, b or c: ", c("a", "b", "c"))
}
}
