% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_v2.R
\name{get_yn_input}
\alias{get_yn_input}
\title{Get yes/no input from user}
\usage{
get_yn_input(prompt)
}
\arguments{
\item{prompt}{The prompt to display to the user.}
}
\value{
TRUE if the user entered 'y' and FALSE if they entered 'n'.
}
\description{
Prompts the user for input until they enter either 'y' or no 'n'. Returns TRUE if the user entered 'y' and FALSE if they entered 'n'.
}
\examples{
\dontrun{
show_dir <- get_yn_input("List dir content? (y/n) ")
if (show_dir) print(dir())
}
}
\keyword{internal}
