% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{headtail}
\alias{headtail}
\title{Show head and tail rows of a matrix-like object}
\usage{
headtail(x, n = 6)
}
\arguments{
\item{x}{A matrix or data frame.}

\item{n}{Number of rows to take from the top and bottom.}
}
\value{
A subset of \code{x} containing head and tail rows.
}
\description{
Returns the first and last \code{n} rows of a matrix or data frame. If the input
has fewer than \code{2*n} rows, overlapping rows are returned only once.
}
\examples{
x <- matrix(seq_len(30), nrow = 10)
headtail(x, n = 2)
}
\author{
2024-2026 Tobias Schmidt: initial version.
}
