% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_v2.R
\name{is.none}
\alias{is.none}
\title{Check if a value is NULL or NA}
\usage{
is.none(x)
}
\arguments{
\item{x}{The value to check.}
}
\value{
TRUE if x is NULL or x is NA, else FALSE.
}
\description{
Checks if a given value is NULL or NA.
}
\keyword{internal}
