% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_spectrum.R
\name{is.spectrum}
\alias{is.spectrum}
\title{Check whether an Object is a Spectrum}
\usage{
is.spectrum(xx, check.class = TRUE, check.contents = FALSE)
}
\description{
An object is considered a 'Spectrum', if it
\itemize{
\item is a list
\item has all 'Mandatory Elements' listed below
\item fulfills all 'Constraints' listed below
}
}
\section{Mandatory Elements}{
\itemize{
\item \code{si}: Measured signal intensities in arbitrary units (au)
\item \code{cs}: Corresponding "chemical shifts" in parts per pillion (ppm)
\item \code{fq}: Corresponding frequencies in Hertz (Hz)
}
}

\section{Optional Elements}{
\itemize{
\item \code{name}: Name of the spectrum, e.g. "Blood 1" or "Urine Mouse X23D"
\item \code{path} are character vectors of length 1, e.g. "/example_datasets/bruker/urine/urine_1"
\item \code{type}: Type of the spectrum, e.g. "H1 CPMG" or "H1 NOESY"
\item \code{mfs}: Magnetic field strength in Tesla
}
}

\section{Constraints}{
\itemize{
\item Elements \code{si}, \code{fq} and \code{cs} must have the same length.
\item Each of the elements \code{name}, \code{type} and \code{path} must be a character vector of length 1 or NULL.
\item Element \code{mfs} must be a numeric vector of length 1 or NULL.
}
}

\examples{
spectrum <- read_spectrum()
is.spectrum(spectrum)
}
