% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_read_spectrum.R
\name{parse_bruker_param_file}
\alias{parse_bruker_param_file}
\title{Parse Bruker Parameter File}
\usage{
parse_bruker_param_file(path)
}
\arguments{
\item{path}{The path of the file containing the parameter data. E.g. \code{"example_datasets/bruker/urine/urine_1/10/acqus"} or \code{"example_datasets/bruker/urine/urine_1/10/pdata/10/procs"}.}
}
\description{
Parses a Bruker parameter file and returns the parameters as a named list.
}
\details{
For a detailed description of the format of burker parameter files, refer to 'Bruker_NMR_Data_Formats.pdf'.
}
