% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{plot.mdlm}
\alias{plot.mdlm}
\title{Plot mdlm object}
\usage{
\method{plot}{mdlm}(x, ...)
}
\arguments{
\item{x}{Fitted \code{mdlm} object.}

\item{...}{Additional plotting arguments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plots the lasso path for \code{mdlm} objects.
}
\examples{
\dontrun{
  m <- cv_mdlm(spectra, y, sfr = c(11, -2))
  plot(m)
}
}
