% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{plot.mdm}
\alias{plot.mdm}
\title{Plot mdm object}
\usage{
\method{plot}{mdm}(x, ...)
}
\arguments{
\item{x}{Fitted \code{mdm} object.}

\item{...}{Additional plotting arguments.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Plots the lasso path for an \code{mdm} object.
}
\examples{
\dontrun{
  m <- fit_mdm_grid(spectra, y, sfr = c(11, -2))
  plot(m)
}
}
