% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/55_plot_aligned_spectra.R
\name{plot_aligned_spectra}
\alias{plot_aligned_spectra}
\title{Plot Aligned Spectra}
\usage{
plot_aligned_spectra(YA, YB, PA, PB, mfcol = c(nrow(YA), 1))
}
\arguments{
\item{YA}{Matrix. YA\link{i,j} == Signal Intensity of spectrum i at index j AFTER alignmnent.}

\item{YB}{Matrix. YB\link{i,j} == Signal Intensity of spectrum i at index j BEFORE alignment.}

\item{PA}{List of vectors. PA[\link{i}][j] == index of peak j of spectrum i AFTER alignment.}

\item{PB}{List of vectors. PB[\link{i}][j] == index of peak j of spectrum i BEFORE alignment.}

\item{mfcol}{Vector of two integers specifying the number of rows and columns of the plot grid}
}
\description{
Plots the aligned and unaligned spectra for comparison.
}
\examples{
x <- seq(1.5 * pi, 9.5 * pi, length.out = 90)
y <- 10 * sin(x) # y without noise
p <- sort(order(y, decreasing = TRUE)[1:4]) # peaks without noise
Y <- lapply(1:4, function(i) smooth(smooth(y + rnorm(90)))) # add noise
Y <- do.call(rbind, Y)
Y <- Y - min(Y)
YB <- rbind(
    c(rep(0, 1), Y[1, ], rep(0, 9)),
    c(rep(0, 4), Y[1, ], rep(0, 6)),
    c(rep(0, 8), Y[1, ], rep(0, 2)),
    c(rep(0, 3), Y[1, ], rep(0, 7))
)
YA <- rbind(
    c(rep(0, 5), Y[1, ], rep(0, 5)),
    c(rep(0, 5), Y[1, ], rep(0, 5)),
    c(rep(0, 5), Y[1, ], rep(0, 5)),
    c(rep(0, 5), Y[1, ], rep(0, 5))
)
PA <- list(p + 5, p + 5, p + 5, p + 5)
PB <- list(p + 1, p + 4, p + 8, p + 3)
plot_aligned_spectra(YA, YB, PA, PB)
}
