% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDecon1D.R
\name{plot_lorentz_curves_save_as_png}
\alias{plot_lorentz_curves_save_as_png}
\title{Plot lorentz curves for variable range}
\usage{
plot_lorentz_curves_save_as_png(deconv_result, x_range = c(), y_range = c())
}
\arguments{
\item{deconv_result}{Saved result of the MetaboDecon1D() function}

\item{x_range}{Row vector with two entries consisting of the ppm start and
the ppm end value to scale the range of the x-axis (optional)}

\item{y_range}{Row vector with two entries consisting of the ppm start and
the ppm end value to scale the range of the y-axis (optional)}
}
\description{
The MetaboDecon1D package also comprise the additional function
plot_lorentz_curves_save_as_png() to plot the original spectrum and all
generated Lorentz curves and save the result as png under the filepath.
}
\examples{
\dontrun{
# Load one spectrum (Bruker format)
result <- MetaboDecon1D(filepath="load_example_path", filename="urine", file_format="bruker")
plot_lorentz_curves_save_as_png(result)

# Load more spectra (Bruker format)
result <- MetaboDecon1D(filepath="load_example_path", file_format="bruker")
# Plot Lorentz curves of all investigated spectra
plot_lorentz_curves_save_as_png(result)
# Plot Lorentz curves of one certain investigated spectrum
plot_lorentz_curves_save_as_png(result$urine)
}
}
\seealso{
\code{\link{MetaboDecon1D}},
\code{\link{plot_triplets}},
\code{\link{plot_spectrum_superposition_save_as_png}}
}
