% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_spectrum_v2.R
\name{plot_sfr}
\alias{plot_sfr}
\title{Plot Signal Free Region}
\usage{
plot_sfr(spec, left_ppm, right_ppm)
}
\arguments{
\item{spec}{A list representing the spectrum as returned by \code{\link[=load_jcampdx_spectrum]{load_jcampdx_spectrum()}} or \code{\link[=load_bruker_spectrum]{load_bruker_spectrum()}}.}

\item{left_ppm}{The left border of the signal free region in ppm.}

\item{right_ppm}{The right border of the signal free region in ppm.}
}
\value{
NULL. Called for side effect of plotting the signal free region.
}
\description{
Draws the signal free region as green vertical lines into the given spectrum.
}
