% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDecon1D.R
\name{plot_triplets}
\alias{plot_triplets}
\title{Function to plot peak triplets for variable range}
\usage{
plot_triplets(deconv_result, x_range = c(), y_range = c())
}
\arguments{
\item{deconv_result}{Saved result of the MetaboDecon1D() function}

\item{x_range}{Row vector with two entries consisting of the ppm start and the ppm end value to scale the range of the x-axis (optional)}

\item{y_range}{Row vector with two entries consisting of the ppm start and the ppm end value to scale the range of the y-axis (optional)}
}
\description{
The MetaboDecon1D package also comprise the additional function
plot_triplets() to plot the peak triplets for each peak.
}
\examples{
\dontrun{
# Load one spectrum (Bruker format)
result <- MetaboDecon1D(filepath="load_example_path", filename="urine", file_format="bruker")
plot_triplets(result)
# Plot part of spectrum
plot_triplets(result, c(1.4,1.3), c(0,20))

# Load all spectra (Bruker format) of choosen folder
result <- MetaboDecon1D(filepath="load_example_path", file_format="bruker")
# Plot triplets of all investigated spectra
plot_triplets(result)
# Plot triplets of one certain investigated spectrum
plot_triplets(result$urine)
}
}
\seealso{
\code{\link{MetaboDecon1D}},
\code{\link{calculate_lorentz_curves}},
\code{\link{plot_lorentz_curves_save_as_png}},
\code{\link{plot_spectrum_superposition_save_as_png}}
}
