% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_plot.R
\name{plot_ws}
\alias{plot_ws}
\title{Plot Water Signal}
\usage{
plot_ws(spec, hwidth_ppm)
}
\arguments{
\item{spec}{A list representing the spec as returned by \code{\link[=load_jcampdx_spectrum]{load_jcampdx_spectrum()}} or \code{\link[=load_bruker_spectrum]{load_bruker_spectrum()}}.}

\item{hwidth_ppm}{The half width of the water signal in ppm.}
}
\value{
NULL. Called for side effect of plotting the water signal.
}
\description{
Draws the water signal as red vertical lines into the given spectrum.
}
