% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wip.R
\name{predict_svm}
\alias{predict_svm}
\title{Predict SVM}
\usage{
predict_svm(model, X)
}
\arguments{
\item{model}{Model from train_svm().}

\item{X}{Feature matrix.}
}
\value{
List with prob (probabilities) and cls (predicted classes 0/1).
}
\description{
Predicts probabilities and classes from trained SVM.
}
\examples{
\dontrun{
m <- train_svm(X, y, 1, 0.01, 3)
p <- predict_svm(m, X_test)
}
}
