% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{print.mdlm}
\alias{print.mdlm}
\title{Print mdlm object}
\usage{
\method{print}{mdlm}(x, ...)
}
\arguments{
\item{x}{Fitted \code{mdlm} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints a compact model summary for \code{mdlm} objects.
}
\examples{
m <- structure(
  list(
      model = NULL, ref = NULL,
      meta = list(
          model = "lasso", nfit = 5, smit = 2, smws = 5,
          delta = 4, maxShift = 50, maxSnap = 10
      )
  ),
  class = "mdlm"
)
print(m)
}
