% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{print.summary.mdlm}
\alias{print.summary.mdlm}
\title{Print summary.mdlm object}
\usage{
\method{print}{summary.mdlm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.mdlm}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints formatted output for objects of class \code{summary.mdlm}.
}
\examples{
s <- structure(
  list(
      model = "lasso", ref_len = 0, smws = 5,
      delta = 4, maxShift = 50, maxSnap = 10
  ),
  class = "summary.mdlm"
)
print(s)
}
