% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_spectrum.R
\name{raw_spectrum}
\alias{raw_spectrum}
\alias{make_spectrum}
\alias{read_spectrum}
\alias{as_raw_spectrum}
\alias{is_spectrum}
\alias{simulate_spectrum}
\alias{print.spectrum}
\title{Raw Spectrum Class}
\usage{
raw_spectrum(
  si,
  cs,
  fq,
  name = NULL,
  path = NULL,
  type = NULL,
  mfs = NULL,
  version = packageVersion("metabodecon")
)

make_spectrum(
  si,
  cs_max,
  cs_width,
  fq_ref,
  fq_width = NULL,
  force = FALSE,
  silent = FALSE,
  name = NULL,
  path = NULL,
  type = NULL,
  mfs = NULL
)

read_spectrum(
  path = metabodecon_file("bruker/sim/sim_01"),
  file_format = "bruker",
  expno = 10,
  procno = 10,
  raw = FALSE,
  silent = TRUE,
  force = FALSE
)

as_raw_spectrum(x, ...)

is_spectrum(x, check_class = TRUE, check_contents = FALSE)

simulate_spectrum(
  deconv = glc("blood_01")$rv$ret,
  show = TRUE,
  pngpath = NULL,
  pdfpath = NULL,
  svgpath = NULL,
  rdspath = NULL,
  brukerdir = NULL,
  verbose = TRUE,
  noise_method = "SFR"
)

\method{print}{spectrum}(x, ...)
}
\arguments{
\item{si}{Signal intensities in Arbitrary Units (au), ordered from highest to lowest corresponding chemical shift.}

\item{cs}{Vector of "chemical shifts" in parts per pillion (ppm). Must be of the same length as \code{si} and \code{fq}.}

\item{fq}{Vector of "frequencies" in Hertz (Hz). Must be of the same length as \code{si} and \code{cs}.}

\item{name}{The name of the spectrum, e.g. "Blood 1" or "Urine Mouse X23D".}

\item{path}{The path of the file/folder containing the spectrum data. E.g. \code{"example_datasets/jcampdx/urine/urine_1.dx"} or \code{"example_datasets/bruker/urine/urine"}.}

\item{type}{The type of experiment, e.g. "H1 CPMG" or "H1 NOESY".}

\item{mfs}{Magnetic field strength in Tesla, e.g. \code{14.1}.}

\item{version}{The version of 'metabodecon' that was used to create the \code{spectrum} object.}

\item{cs_max}{The highest chemical shift value in ppm, usually shown as left end of the spectrum.}

\item{cs_width}{The width of the spectrum in ppm.}

\item{fq_ref}{The frequency of the reference in Hz.}

\item{fq_width}{The width of the spectrum in Hz. If provided, the values calculated from \code{cs_max}, \code{cs_width} and \code{fq_ref} are compared against this value. If the values differs, an error is thrown. If set to \code{NULL}, this check will be skipped.}

\item{force}{If \code{TRUE}, try to continue when encountering errors and print info messages instead. To hide these messages as well, set \code{silent = TRUE}.}

\item{silent}{If \code{TRUE}, no output will be printed to the console.}

\item{file_format}{The file format of the spectrum file. E.g. \code{"bruker"} or \code{"jcampdx"}.}

\item{expno}{The experiment number for the file. E.g. \code{"10"}. Only relevant if \code{file_format} equals \code{"bruker"}. For details about \code{expno} and \code{procno} see section \href{https://spang-lab.github.io/metabodecon/articles/FAQ.html#file-structure}{File Structure} in the metabodecon FAQ.}

\item{procno}{The processing number for the file. E.g. \code{"10"}. Only relevant if \code{file_format} equals \code{"bruker"}.}

\item{raw}{If \code{FALSE}, scales the returned signal intensities based on information available in the spectrum metadata, in particular \code{NC_proc}. For details see \code{processing-reference.pdf}, available at \url{https://www.bruker.com/en.html} at section 'Services & Support > Documentation & Manuals > Magnetic Resonance > Acquisition & Processing > TopSpin Processing Commands and Parameters' (requires login).}

\item{x}{A \link{raw_spectrum} object.}

\item{check_class}{If TRUE, class attribute "raw_spectrum" is required or the check will fail.}

\item{check_contents}{If TRUE, the contents of the provided object must comply to the expected structure of \code{raw_spectrum} objects, described in 'Details' or the check will fail.}

\item{deconv}{A list containing the deconvolution result, as returned by \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}.}

\item{show}{Logical indicating whether to display the simulated spectrum.}

\item{pngpath}{Path to save the simulated spectrum as a PNG file.}

\item{pdfpath}{Path to save the simulated spectrum as a PDF file.}

\item{svgpath}{Path to save the simulated spectrum as a SVG file.}

\item{rdspath}{Path to save the simulated spectrum as an RDS file.}

\item{brukerdir}{Path to the directory where the Bruker files should be saved.}

\item{verbose}{Logical indicating whether to print progress messages.}

\item{noise_method}{Method to generate noise. Either \code{"RND"} or \code{"SFR"}.}
}
\value{
A \code{raw_spectrum} object as described in 'Details', except for \code{is_spectrum()}, which returns a locical.
}
\description{
Objects of class \code{raw_spectrum} are used by 'metabodecon' to represent NMR spectra that have not been devonvoluted yet. The following functions can be used to create \code{raw_spectrum} objects:
\itemize{
\item \code{raw_spectrum()}: Create from scratch by specifying all elements directly. Usually not used directly, but through one of the following functions.
\item \code{make_spectrum()}: Create by specifying signal intensities, chemical shifts and reference frequency.
\item \code{read_spectrum()}: Create from local file or folder.
\item \code{simulate_spectrum()}: Create from existing deconvolution result.
\item \code{as_raw_spectrum()}: Create from existing compatible object.
}

To test for \code{raw_spectrum} objects, \code{is_spectrum()} can be used.
}
\details{
A \code{raw_spectrum} object is a list with class attribute \code{"spectrum"} and the following elements in arbitrary order: \code{si}, \code{cs}, \code{fq}, \code{name}, \code{path}, \code{type}, \code{mfs}, \code{version}. Each element is described in 'Arguments' and must fulfill the following requirements:
\itemize{
\item \code{si}, \code{fq} and \code{cs} are numeric vectors of the same length.
\item \code{name}, \code{type} and \code{path} are character vectors of length 1 or NULL.
\item \code{version} is a character vector of length 1 of the form \code{x.y} where \code{x} and \code{y} are integers.
\item \code{mfs} is a numeric vector of length 1 or NULL.
}

Element access should always be done by name, e.g. using \code{spectrum_obj$si}, \code{spectrum_obj[["cs"]]} or \code{spectrum_obj[c("fq", "name")]}, because the order of elements is not guaranteed.
}
\examples{

## Create from Scratch

si <- c(1, 1, 3, 7, 8, 3, 8, 5, 2, 1)
cs_max <- 14.8
cs_width <- 20.0
fq_ref <- 600.25 * 1e6
fq_width <- 12005
spectrum1 <- make_spectrum(si, cs_max, cs_width, fq_ref, fq_width)
spectrum2 <- make_spectrum(si, cs_max, cs_width, fq_ref, fq_width = 12010, force = FALSE)
print(spectrum1)
print(spectrum2)
is_spectrum(spectrum1)


## Read from Disk

spectrum3 <- read_spectrum(metabodecon_file("urine_1"))
print(spectrum3)

\dontrun{
# Reading files in JCAMP-DX format is very slow (about 30s on the development
# machine). So if possible, you should stick with the original Bruker format.
urine_1_dx <- system.file("example_datasets/jcampdx/urine/urine_1.dx", package = "metabodecon")
spectrum4 <- read_spectrum(urine_1_dx, file_format = "jcampdx")
stopifnot(all.equal(urine_1_spectrum, spectrum4))
}


## Simulate

spectrum5 <- simulate_spectrum()
print(spectrum5)

\dontrun{
     deconv <- glc("blood_02")$rv$ret
     show = FALSE
     pngpath <- "vignettes/Datasets/png/sim_02.png"
     pdfpath <- "vignettes/Datasets/pdf/sim_02.pdf"
     svgpath <- "vignettes/Datasets/svg/sim_02.svg"
     rdspath <- "inst/example_datasets/rds/sim/sim_02.rds"
     brukerdir <- "inst/example_datasets/bruker/sim/sim_02"
     spectrum6 <- simulate_spectrum(deconv, show, pngpath, pdfpath, svgpath, rdspath, brukerdir)
}
}
