% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_classes.R
\name{RawSpectra}
\alias{RawSpectra}
\title{RawSpectra Class}
\usage{
RawSpectra(...)
}
\description{
Each object of this class:
\itemize{
\item Represents one or more one dimensional NMR spectra.
\item Is a list of \link{spectrum} objects.
}

Objects of this class can be created using \code{\link[=read_spectra]{read_spectra()}}.
}
\examples{
fq <- c(0.599771, 0.691471, 0.783171, 0.874871, 0.966571) + 600248506
cs <- c(7.164231, 7.164079, 7.163926, 7.163773, 7.163620)
si <- c(100861.5, 105161.5, 109285.3, 113199.5, 116719.7)
si1 <- si + rnorm(5, 0, 100)
si2 <- si + rnorm(5, 0, 100)
raw_spectrum1 <- spectrum(si1, cs, fq)
raw_spectrum2 <- spectrum(si2, cs, fq)
raw_spectra <- RawSpectra(raw_spectrum1, raw_spectrum2)
try(RawSpectra(raw_spectrum1, 2))
try(RawSpectra(raw_spectrum1, 2, raw_spectrum2, "a", "b"))
}
