% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_spectrum_v2.R
\name{read_1r_file}
\alias{read_1r_file}
\title{Read Bruker TopSpin spectrum from 1r file}
\usage{
read_1r_file(path, expno, procno, procs)
}
\arguments{
\item{path}{The path of the directory holding the spectrum data. E.g. \code{"example_datasets/bruker/urine/urine_1/"}.}

\item{expno}{The experiment number for the file. E.g. \code{"10"}.}

\item{procno}{The processing number for the file. E.g. \code{"10"}.}

\item{procs}{The content of the \code{procs} file. E.g. \code{readLines(file.path(path, expno, "pdata", procno, "procs"))}.}
}
\value{
The signals read from the file as numeric vector
}
\description{
Read Bruker TopSpin spectrum from 1r file
}
\examples{
\dontrun{
xds_path <- download_example_datasets()
path <- file.path(xds_path, "bruker/urine/urine_1")
procs <- readLines(file.path(path, 10, "pdata", 10, "procs"))
y <- read_1r_file(path, 10, 10, procs)
}
}
