% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock.R
\name{redirect}
\alias{redirect}
\title{Redirect output, message, and plot streams}
\usage{
redirect(output = "captured", message = NULL, plots = NULL)
}
\arguments{
\item{output}{The target for the output stream. Defaults to "captured".}

\item{message}{The target for the message stream. Defaults to NULL.}

\item{plots}{The target for the plot stream. Defaults to NULL. If a file path is provided, plots will be saved to a PDF at that location.}
}
\value{
A list of environments, one for each stream, containing the redirection settings.
}
\description{
Redirects the output, message, and plot streams in R to either a specified file or a captured character vector.
If the target is "captured", the stream is captured into a character vector.
If the target is NULL, the stream is not redirected.
Otherwise, the target is assumed to be a file path, and the stream is redirected to that file.
}
\examples{
# Capture output

redirects <- redirect(output = "captured")
cat("Hello\n")
cat("World\n")
restore()
ls.str(redirects$output)

# Capture messages, redirect output to a file, and save plots to a PDF

redirects <- redirect(output = "output.txt", message = "captured", plots = "plots.pdf")
print("This goes to output.txt")
message("This is captured")
warning("This is captured as well")
plot(1:10) # This plot is saved to plots.pdf
restore()
cat(redirects$message$text)
}
