% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sap}
\alias{sap}
\title{The SAP Dataset}
\format{
An object of class \code{spectra} of length 1.
}
\usage{
sap
}
\description{
The SAP Dataset consists of a single 'Simple-As-Possible' (SAP) spectrum. The
purpose of the SAP spectrum is to provide a straightforward example that can
be used to test and understand the deconvolution algorithm in detail.
}
\details{
The first (and only) spectrum within the SAP dataset contains 128 datapoints
ranging from -6.4 to 6.4 ppm with four peaks. A rough sketch of the spectrum
is shown below:

\if{html}{\out{<div class="sourceCode">}}\preformatted{-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~-~
|      SFR      |               w               |     SFR      |
|               |  x           www       p      |              |
|~-~-~-~-~-~-~-~|~-|-|-~-~-~-~-~|~-~-~-~-|-~-~-~-~-~-~-~-~-~-~-~
|               |  | |          |        |      |
6.4             |  | 2.24       0.047    -2.22  -3.2
                |  2.61
                3.2
}\if{html}{\out{</div>}}
}
\keyword{datasets}
