% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflection_points_v2.R
\name{select_inflection_points_v0}
\alias{select_inflection_points_v0}
\title{Select Inflection Points}
\usage{
select_inflection_points_v0(spectrum)
}
\arguments{
\item{spectrum}{A list with elements \code{sdp} and \code{ssss} representing the x-values in scaled data points and the y values in "smoothed and scaled signal strength" respectively.}

\item{details}{this is the code copy pasted from the original \code{deconvolution} function of \code{MetaboDecon1D} with only the inputs renamed to \code{spectrum$sdp} and \code{spectrum$ssss}.}
}
\description{
This function calculates the second derivative of a spectrum and finds all local minima of the derivate, i.e. inflection points. It returns a list containing the second derivative, the x-values of the local minima, and their indices.
}
