% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflection_points_v2.R
\name{select_inflection_points_v1}
\alias{select_inflection_points_v1}
\title{Select Inflection Points}
\usage{
select_inflection_points_v1(spectrum)
}
\arguments{
\item{spectrum}{A list with elements \code{sdp} and \code{ssss} representing the x-values in scaled data points and the y values in "smoothed and scaled signal strength" respectively.}

\item{details}{This is the same as \code{select_inflection_points_v0}, but using different variable names within the code to make it more readable.}
}
\description{
This function calculates the second derivative of a spectrum and finds all local minima of the derivate, i.e. inflection points. It returns a list containing the second derivative, the x-values of the local minima, and their indices.
}
