% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R, R/rust.R
\name{set_up_deconvoluter}
\alias{set_up_deconvoluter}
\title{Set Up Deconvoluter}
\usage{
set_up_deconvoluter(nfit, smopts, delta, ignore_regions)

set_up_deconvoluter(nfit, smopts, delta, ignore_regions)
}
\arguments{
\item{nfit}{Number of iterations for the analytical fitter.}

\item{smopts}{Smoothing options as a vector of two elements: iterations and window size.}

\item{delta}{Noise score threshold.}

\item{ignore_regions}{Regions to ignore during deconvolution.}
}
\value{
A Deconvoluter object.

A Deconvoluter object.
}
\description{
Sets up the deconvoluter with specified parameters.

Sets up the deconvoluter with specified parameters.
}
