% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_old_metabodecon1d.R
\name{show_license}
\alias{show_license}
\title{Show terms and conditions of license}
\usage{
show_license(...)
}
\arguments{
\item{...}{Not used}
}
\description{
Show terms and conditions of license
}
