% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{The Sim Dataset}
\format{
A \code{spectra} object consisting of 16 \code{spectrum} objects, where each spectrum
contains 2048 datapoints ranging from 3.60 to 3.29 ppm. For details about
\code{spectrum} and \code{spectra} objects see \link{metabodecon-classes}.
}
\usage{
sim
}
\description{
A simulated dataset generated from the \href{https://spang-lab.github.io/metabodecon/articles/Datasets.html#Blood}{Blood}
dataset.
}
\keyword{datasets}
