% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_spectra.R
\name{spectra}
\alias{spectra}
\title{Spectra Class}
\usage{
spectra(...)
}
\description{
Each object of this class is a list of \code{spectrum} objects with class attribute \code{spectra}.
}
\examples{
fq <- c(0.599771, 0.691471, 0.783171, 0.874871, 0.966571) + 600248506
cs <- c(7.164231, 7.164079, 7.163926, 7.163773, 7.163620)
si <- c(100861.5, 105161.5, 109285.3, 113199.5, 116719.7)
si1 <- si + rnorm(5, 0, 100)
si2 <- si + rnorm(5, 0, 100)
spectrum1 <- spectrum(si1, cs, fq)
spectrum2 <- spectrum(si2, cs, fq)
spectra <- spectra(spectrum1, spectrum2)
try(spectra(spectrum1, 2))
try(spectra(spectrum1, 2, spectrum2, "a", "b"))
}
