% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdlm.R
\name{summary.mdlm}
\alias{summary.mdlm}
\title{Summarize mdlm object}
\usage{
\method{summary}{mdlm}(object, ...)
}
\arguments{
\item{object}{Fitted \code{mdlm} object.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Object of class \code{summary.mdlm}.
}
\description{
Builds a compact summary list for mdlm objects.
}
\examples{
m <- structure(
  list(
      ref = NULL,
      meta = list(
          model = "lasso", nfit = 5, smit = 2, smws = 5,
          delta = 4, maxShift = 50, maxSnap = 10
      )
  ),
  class = "mdlm"
)
summary(m)
}
