% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadir.R
\name{tempdir}
\alias{tempdir}
\title{Temporary Session Directory}
\usage{
tempdir()
}
\value{
Returns the path to the temporary session directory.
}
\description{
Returns the path to metabodecon's temporary session directory.
This directory equals subdirectory \code{"metabodecon"} of R's temporary session directory \verb{[tempdir()]} plus additional path normalization.
}
