% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_datadir.R
\name{tmpdir}
\alias{tmpdir}
\title{Temporary Session Directory}
\usage{
tmpdir()
}
\value{
Returns the path to the temporary session directory.
}
\description{
Returns the path to metabodecon's temporary session directory.
This directory equals subdirectory \code{"metabodecon"} of R's temporary session directory \verb{[base::tempdir()]} plus additional path normalization.
}
