% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_util.R
\name{transp}
\alias{transp}
\title{Make transparent}
\usage{
transp(col = "violet", alpha = 0.08)
}
\arguments{
\item{col}{Character string specifying the color to make transparent.}

\item{alpha}{Numeric value between 0 and 1 specifying the transparency level.}
}
\value{
A character string representing the color with an alpha channel.
}
\description{
Make a color transparent by adding an alpha channel.
}
\examples{
transp("violet", 0.08)
transp("black", 0.5)
}
