% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_is.R
\name{type_checking}
\alias{type_checking}
\alias{type}
\alias{is_spectrum}
\alias{is_gspec}
\alias{is_gdecon}
\alias{is_decon0}
\alias{is_decon1}
\alias{is_decon2}
\alias{is_spectra}
\alias{is_gspecs}
\alias{is_gdecons}
\alias{is_decons0}
\alias{is_decons1}
\alias{is_decons2}
\title{Check the type of an object}
\usage{
type(x)

is_spectrum(x, check_class = TRUE, check_contents = FALSE)

is_gspec(x)

is_gdecon(x)

is_decon0(x)

is_decon1(x)

is_decon2(x)

is_spectra(
  x,
  check_class = TRUE,
  check_contents = FALSE,
  check_child_classes = FALSE
)

is_gspecs(x)

is_gdecons(x)

is_decons0(x)

is_decons1(x)

is_decons2(x)
}
\arguments{
\item{x}{The object to check.}

\item{check_class}{Logical indicating whether to check the class of the object.}

\item{check_contents}{Logical indicating whether to check the contents of the object.}

\item{check_child_classes}{Logical indicating whether to check the class of each element of the object.}
}
\value{
For \code{type()}, a character string indicating the type of the object. For \verb{is_*} functions, a logical indicating whether the object is of the specified type.
}
\description{
Functions starting with \verb{is_*} check objects for a specific type. Function \code{type()} is similar to \code{typeof} but also returns the length of the object for logical, integer, double, complex, character, and raw vectors. For other objects, it returns the class of the object.
}
\examples{
obj <- structure(list(a="a", b=1:5), class = "abc")
type(obj)   # "abc"
type(1:5)   # "integer(5)"
type(NULL)  # "NULL(0)"
type("Hi")  # "character(1)"
type(1.0)   # "numeric(1)"
type(NA)    # "logical(1)"
type(NaN)   # "numeric(2)"
}
