% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/90_test_helpers.R
\name{vcomp}
\alias{vcomp}
\title{Compare two vectors}
\usage{
vcomp(x, y, xpct = 0, silent = FALSE)
}
\value{
0==identical, 1==all.equal, 2==different, 3==error
}
\description{
Checks if \code{x} and \code{y} are identical, all.equal or different vectors. If different, differences are printed.
}
\examples{
x <- 1:10
y1 <- 1:10
y2 <- c(1:4, 5.000000000000001, 6:8, 8.99999999999999, 10)
y3 <- c(1:4, 5.1, 6:8, 8.9, 10)
y4 <- list("a", "b", "c")
vcomp(x, y1)
vcomp(x, y2)
vcomp(x, y3)
vcomp(x, y4)
}
