% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_convert.R
\name{width}
\alias{width}
\title{Calculate the Width of a Numeric Vector}
\usage{
width(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The width of the vector, calculated as the difference between its maximum and
minimum values.
}
\description{
Calculates the width of a numeric vector by computing the difference between
the maximum and minimum values in the vector.
}
\examples{
vec <- c(1, 3, 5, 7, 9)
width(vec)
}
