% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{Align Spectra}
\usage{
align(x, ref = NULL, maxShift = 50, verbose = TRUE, nworkers = 1)
}
\arguments{
\item{x}{An object of type \code{decons2} or \code{aligns}, as described in \link{metabodecon-classes}.}

\item{ref}{Optional reference spectrum of type \code{align} or \code{decon2}. When supplied,
all spectra in \code{x} are aligned towards this reference. The reference is
prepended to \code{x} internally and removed from the result. If \code{NULL}
(default), the reference is chosen automatically.}

\item{maxShift}{Maximum number of datapoints a peak center may be shifted during CluPA
alignment. 50 is a suitable starting value for plasma spectra with a digital
resolution of 128K. Increase for urine or other sample types with larger
chemical-shift variation. Use \code{maxShift = 0L} to skip alignment entirely:
each peak's aligned center \code{x0al} is set equal to its fitted center \code{x0}.}

\item{verbose}{Whether to print progress messages during alignment.}

\item{nworkers}{Number of parallel workers. Default is 1 (no parallelism).}
}
\value{
An object of type \code{aligns} as described in \link{metabodecon-classes}.
}
\description{
Align signals across  a  list  of  deconvoluted  spectra  using  the  'CluPA'
algorithm from the 'speaq' package, described  in  Beirnaert  et  al.  (2018)
\url{doi:10.1371/journal.pcbi.1006018}     and     Vu     et      al.      (2011)
\url{doi:10.1186/1471-2105-12-405} plus the additional peak combination described
in \code{\link[=combine_peaks]{combine_peaks()}}.
}
\examples{
decons <- deconvolute(sim[1:2], sfr = c(3.55, 3.35))
aligned <- align(decons)
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
