% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cache_example_datasets}
\alias{cache_example_datasets}
\title{Cache Example Files}
\usage{
cache_example_datasets(persistent = FALSE, overwrite = FALSE)
}
\arguments{
\item{persistent}{Logical. If TRUE, the datasets are cached permanently. If NULL, the function prompts the user. Default is NULL.}
}
\value{
The path to the cached datasets.
}
\description{
Simpler Version of \code{\link[=cache_example_datasets]{cache_example_datasets()}} that does not prompt the user for input.
}
\examples{
cache_example_datasets(persistent = FALSE)
}
