% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_old_metabodecon1d.R
\name{calculate_lorentz_curves}
\alias{calculate_lorentz_curves}
\title{Calculate lorentz curves for each analyzed spectrum}
\usage{
calculate_lorentz_curves(deconv_result, number_of_files = NA)
}
\arguments{
\item{deconv_result}{Saved result of the MetaboDecon1D() function}

\item{number_of_files}{Number of spectra to analyze}
}
\value{
\bold{One spectrum:} matrix containing the generated Lorentz curves. Each row depicts one Lorentz curve. \bold{More spectra:} list containing the matrix with the generated Lorentz curve. Each matrix corresponds to one spectrum and each row in the matrix depicts one Lorentz curve.
}
\description{
Calculates the lorentz curves of each investigated spectrum.
}
\examples{
\dontrun{

# Load one spectrum (Bruker format)
result <- MetaboDecon1D(
  filepath = "load_example_path",
  filename = "urine",
  file_format = "bruker"
)
lorentz_curves <- calculate_lorentz_curves(result)

# Load all spectra (Bruker format) of choosen folder
result <- MetaboDecon1D(filepath = "load_example_path", file_format = "bruker")

# Calculate Lorentz curves of all investigated spectra
lorentz_curves <- calculate_lorentz_curves(result)

# Save matrix containing all Lorentz curves of first spectrum
lorentz_curves_first_spectrum <- lorentz_curves[[1]]

# Calculate Lorentz curves of one certain investigated spectrum
lorentz_curves_spectrum1 <- calculate_lorentz_curves(result$urine)

}
}
\seealso{
\code{\link[=MetaboDecon1D]{MetaboDecon1D()}}, \code{\link[=plot_triplets]{plot_triplets()}}, \code{\link[=plot_lorentz_curves_save_as_png]{plot_lorentz_curves_save_as_png()}}, \code{\link[=plot_spectrum_superposition_save_as_png]{plot_spectrum_superposition_save_as_png()}}
}
