% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/31_get_user_input.R
\name{confirm_wshw}
\alias{confirm_wshw}
\title{Confirm water signal half width (WSHW)}
\usage{
confirm_wshw(spec, wshw)
}
\arguments{
\item{spec}{A list representing the spectrum.}

\item{wshw}{The initial half width in ppm.}
}
\value{
The confirmed half width of the water artefact in ppm units.
}
\description{
Repeatly asks the user to refine and/or confirm the half width of the water artefact.
}
