% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_util.R
\name{convert_width}
\alias{convert_width}
\title{Convert values from unit A to unit B}
\usage{
convert_width(xa, wa, wb)
}
\arguments{
\item{xa}{A numeric vector with values in unit A.}

\item{wa}{Width of a certain interval (e.g. the spectrum width) in unit A.}

\item{wb}{With of the same interval in unit B.}
}
\value{
A numeric vector of values converted from unit A to unit B.
}
\description{
Converts values from unit A to unit B using a conversion factor.
}
\examples{
\donttest{
urine_1 <- pkg_file("example_datasets/bruker/urine/urine_1")
deconv <- generate_lorentz_curves(urine_1, ask = FALSE)[[1]]
sdp <- deconv$x_values
sdp_width <- diff(range(sdp))
ppm <- deconv$x_values_ppm
ppm_width <- diff(range(ppm))
lambda_sdp <- deconv$lambda
lambda_ppm <- convert_width(lambda_sdp, sdp_width, ppm_width)
}
}
