% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{cv_mdm}
\alias{cv_mdm}
\title{Deconvolute, Align and Classify Spectra in Cross-Validation}
\usage{
cv_mdm(
  spectra,
  y,
  sfr,
  use_rust = TRUE,
  npmax = NULL,
  maxShift = seq(50, 250, 50),
  maxSnap = seq(0.5, 2, 0.5),
  nworkers = 1,
  verbose = TRUE,
  cadir = decon_cachedir()
)
}
\arguments{
\item{spectra}{List-like spectra object with \code{cs} and \code{si} vectors.}

\item{y}{Factor vector with class labels for each spectrum.}

\item{sfr}{Signal free region. See \code{\link[=deconvolute]{deconvolute()}} for details.}

\item{use_rust}{Logical. Whether to use the Rust backend.}

\item{nworkers}{Number of workers for parallel grid deconvolution.}

\item{verbose}{Logical. Whether to print log messages.}

\item{cadir}{Directory for caching deconvolution grid search results.
Defaults to \code{\link[=decon_cachedir]{decon_cachedir()}}. If \code{NULL}, caching is disabled.
Pass a custom path to use a different cache directory.}

\item{nfolds}{Number of folds for inner cv.glmnet. Default 10.}
}
\value{
A object of class \code{mdm} as returned by \code{\link[=fit_mdm]{fit_mdm()}}, with two
additional elements \code{pgrid} and \code{ibest}, containing the performance grid and
index of the best parameter combination.
}
\description{
Performs deconvolution, alignment and peak-snapping over a grid of parameters
and calls \code{\link[glmnet:cv.glmnet]{glmnet::cv.glmnet()}} on each corresponding matrix. Calculates the
corresponding \code{cv.glmnet} performances. The deconvolution, alignment and
snapping parameters that enable the best performances are considered optimal
and are used for fitting the final model with the corresponding optimal
lambda value.
}
\examples{
\dontrun{
     aki <- read_aki_data()
     spectra <- aki$spectra
     y <- factor(aki$meta$type, levels=c("Control", "AKI"))
     nworkers <- ceiling(parallel::detectCores() / 2)
     stub(cv_mdm, spectra=spectra, y=y, sfr=c(11, -2), nworkers=nworkers)
     mdm <- cv_mdm(spectra, y, sfr=c(11, -2))
}
}
