% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaboDecon1D.R
\name{deconvolution}
\alias{deconvolution}
\title{Deconvolute one single spectrum}
\usage{
deconvolution(
  filepath,
  name,
  file_format,
  same_parameter,
  processing_value,
  number_iterations,
  range_water_signal_ppm,
  signal_free_region,
  smoothing_param,
  delta,
  scale_factor,
  current_filenumber
)
}
\arguments{
\item{filepath}{TODO}

\item{name}{TODO}

\item{file_format}{TODO}

\item{same_parameter}{TODO}

\item{processing_value}{TODO}

\item{number_iterations}{TODO}

\item{range_water_signal_ppm}{TODO}

\item{signal_free_region}{TODO}

\item{smoothing_param}{TODO}

\item{delta}{TODO}

\item{scale_factor}{TODO}

\item{current_filenumber}{TODO}
}
\value{
TODO
}
\description{
Deconvolute one single spectrum
}
