% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_example_datasets}
\alias{download_example_datasets}
\title{Download metabodecon Example Datasets}
\usage{
download_example_datasets(
  dst_dir = NULL,
  extract = FALSE,
  cache_persistent = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{dst_dir}{The destination directory where the downloaded datasets will be stored. If NULL, the function will return the path to the cached zip file.}

\item{extract}{Logical. If TRUE, the downloaded zip file will be extracted.}

\item{cache_persistent}{Logical. If TRUE, the downloaded datasets will be cached for persistent use.}

\item{overwrite}{Logical. If TRUE, existing files with the same name in the destination directory will be overwritten.}
}
\value{
The path to the downloaded (and possibly extracted) datasets.
}
\description{
This function downloads example datasets that can be used to test the functionality of the metabodecon package.
These datasets are not included in the package by default due to size constraints. The datasets are downloaded as a zip file.
}
\examples{
\dontrun{
download_example_datasets(dst_dir = ".", extract = TRUE, cache_persistent = TRUE, overwrite = TRUE)
}
}
\seealso{
\code{\link[=datadir]{datadir()}}
}
