% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{download_example_datasets}
\alias{download_example_datasets}
\title{Download metabodecon example datasets}
\usage{
download_example_datasets()
}
\description{
Downloads the example files shipped wit metabodecon.  Due to the size
constraints for R packages, these datasets are not included by default when
the package is installed, but must be downloaded explicitly afterwards.
}
\examples{
\dontrun{
download_example_datasets()
}
}
\seealso{
\code{\link[=get_data_dir]{get_data_dir()}}
}
