% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_download_example_datasets.R
\name{download_example_datasets}
\alias{download_example_datasets}
\title{Download metabodecon Example Datasets}
\usage{
download_example_datasets(
  dst_dir = NULL,
  extract = TRUE,
  persistent = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{dst_dir}{The destination directory where the downloaded datasets will be stored. If NULL, the function will return the path to the cached zip file.}

\item{extract}{Logical. If TRUE, the downloaded zip file will be extracted.}

\item{persistent}{Logical. If TRUE, the downloaded datasets will be cached at \code{\link[=datadir_persistent]{datadir_persistent()}} to speed up future calls to download_example_datasets().
If FALSE, the datasets will be cached at \code{\link[=datadir_temp]{datadir_temp()}}.
If NULL, the function will check both paths for the cached datasets but will use \code{\link[=datadir_temp]{datadir_temp()}} if the cached file does not yet exist.}

\item{overwrite}{Logical. If TRUE, existing files with the same name in the destination directory will be overwritten.}
}
\value{
The path to the downloaded (and possibly extracted) datasets.
}
\description{
This function downloads example datasets that can be used to test the functionality of the metabodecon package.
These datasets are not included in the package by default due to size constraints.
The datasets are downloaded as a zip file.
}
\examples{
\dontrun{
download_example_datasets(dst_dir = ".", extract = TRUE, persistent = TRUE, overwrite = TRUE)
}
}
\seealso{
\code{\link[=datadir]{datadir()}}
}
