% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_lorentz_curves}
\alias{generate_lorentz_curves}
\title{Generate Lorentz Curves from NMR Spectra}
\usage{
generate_lorentz_curves(
  data_path,
  file_format = c("bruker", "jcampdx"),
  make_rds = FALSE,
  file_name = NULL,
  number_iterations = 10,
  range_water_signal_ppm = 0.1527692,
  signal_free_region = c(11.44494, -1.8828),
  smoothing_param = c(2, 5),
  delta = 6.4,
  scale_factor = c(1000, 1e+06),
  ask = TRUE
)
}
\arguments{
\item{data_path}{(string) Path to the folder where the original spectra are stored. After deconvolution this folder contains two additional .txt files for each spectrum which contain the spectrum approximated from all deconvoluted signals and a parameter file that contains all numerical values of the deconvolution.}

\item{file_format}{(string) Format of the spectra files.}

\item{make_rds}{(bool) Store results as a rds file on disk? Should be set to TRUE if many spectra are evaluated to decrease computation time.}

\item{file_name}{(string) Name of the NMR file in case only one specific spectrum within \code{data_path} should be analyzed. Note: for Bruker format the file name is usually the same as the folder name.}

\item{number_iterations}{(int) Number of iterations for the approximation of the parameters for the Lorentz curves.}

\item{range_water_signal_ppm}{(float) Half width of the water artefact in ppm.}

\item{signal_free_region}{(float) Row vector with two entries consisting of the ppm positions for the left and right border of the signal free region of the spectrum.}

\item{smoothing_param}{(int) Row vector with two entries consisting of the number of smoothing repeats for the whole spectrum and the number of data points (uneven) for the mean calculation.}

\item{delta}{(float) Threshold value to distinguish between signal and noise.}

\item{scale_factor}{(int) Row vector with two entries consisting of the factor to scale the x-axis and the factor to scale the y-axis.}

\item{ask}{(bool) Whether to ask for user input during the deconvolution process. If set to FALSE, the provided default values will be used.}
}
\description{
Deconvolutes NMR spetra and generates a Lorentz curve for each detected signal within a spectra.
}
\details{
First, an automated curvature based signal selection is performed. Each signal is represented by 3 data points to allow the determination of initial Lorentz curves. These Lorentz curves are then iteratively adjusted to optimally approximate the measured spectrum. For each spectrum two text files will be created in the parent folder i.e. the folder given in data path. The spectrum approximated from all deconvoluted signals and a parameter file that contains all numerical values of the deconvolution. Furthermore, the numerical values of the deconvolution are also stored in a data_frame.
}
\examples{
\dontrun{
spectrum_data <- generate_lorentz_curves(
  data_path = system.file(package = "metabodecon"),
  file_format = "jcampdx",
  file_name = "urine.dx"
)
}
}
