% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_datadir.R
\name{get_data_dir}
\alias{get_data_dir}
\title{Retrieve directory path of an example dataset}
\usage{
get_data_dir(dataset_name = c("", "blood", "test", "urine"), warn = TRUE)
}
\arguments{
\item{dataset_name}{Either \code{""}, \code{"test"}, \code{"blood"} or \code{"urine"}.}

\item{warn}{Whether to print a warning message when the example folders do not yet exist, i.e. \code{\link[=download_example_datasets]{download_example_datasets()}} has not been called yet.}
}
\value{
Path to the directory storing the example files.
}
\description{
Returns the path to the directory storing the example files shipped with metabodecon.

Deprecated since metabodecon v1.2.0. Please use \code{\link[=datadir]{datadir()}} instead. See examples below for usage.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
x <- get_data_dir("urine")                     # Deprecated
y <- datadir("example_datasets/bruker/urine")  # Preferred
cat(x, y, sep = "\n")
}
\seealso{
\code{\link[=download_example_datasets]{download_example_datasets()}}
}
