% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get_ppm_range}
\alias{get_ppm_range}
\title{get_ppm_range}
\usage{
get_ppm_range(spectrum_data)
}
\arguments{
\item{spectrum_data}{Data frame generated from the function
generate_lorentz_curves, contains results from the deconvolution of your
spectra}
}
\description{
function to get ppm range of your spectra, this is required for
spectra alignment following deconvolution
}
\details{
determines after deconvolution the signal with the highest and the
signal with the lowest ppm value.
}
\examples{
\dontrun{
ppm_range <- get_ppm_range(spectrum_data = spectrum_data)
}
}
