% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_alignment.R
\name{get_ppm_range}
\alias{get_ppm_range}
\title{Get PPM Range covered by Spectra}
\usage{
get_ppm_range(
  spectrum_data = glc_sim(),
  show = FALSE,
  mar = c(4.1, 4.1, 1.1, 0.1)
)
}
\arguments{
\item{spectrum_data}{A list of deconvoluted spectra as returned by \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}.}

\item{show}{Whether to plot the ppm range on the spectrum plot.}

\item{mar}{The margins of the plot. Passed on to \code{par()}.}
}
\value{
A vector containing the lowest and highest ppm value over all peaks of the provided deconvoluted spectra.
}
\description{
Returns the ppm range across all peaks of the provided deconvoluted spectra.
}
\examples{
spectrum_data = glc_sim()
ppm_rng <- get_ppm_range(spectrum_data, show = TRUE)
print(ppm_rng)
}
\author{
Initial version from Wolfram Gronwald. Refactored by Tobias Schmidt in 2024.
}
