% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_alignment.R
\name{get_ppm_range}
\alias{get_ppm_range}
\title{get_ppm_range}
\usage{
get_ppm_range(spectrum_data)
}
\arguments{
\item{spectrum_data}{Output of \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}}
}
\description{
Get ppm range of your spectra
}
\details{
determines after deconvolution the signal with the highest and the signal with the lowest ppm value.
}
\examples{
\donttest{ppm_range <- get_ppm_range(spectrum_data = spectrum_data)}
}
