% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_alignment.R
\name{get_ppm_range}
\alias{get_ppm_range}
\title{Get ppm range}
\usage{
get_ppm_range(spectrum_data)
}
\arguments{
\item{spectrum_data}{Output of \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}}
}
\value{
A vector containing the lowest and highest ppm value of the provided spectrum.
}
\description{
Returns the ppm range of the provided spectrum.
}
\details{
determines after deconvolution the signal with the highest and the signal with the lowest ppm value.
}
\examples{
\donttest{
xp <- download_example_datasets()
dp <- file.path(xp, "bruker/urine")
ff <- "bruker"
x <- generate_lorentz_curves(dp, ff, ask = FALSE, nfit = 3)
ppm_range <- get_ppm_range(spectrum_data = x)
}
}
