% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simat.R
\name{get_si_mat}
\alias{get_si_mat}
\title{Extract Matrix of aligned Signal Intensities}
\usage{
get_si_mat(x, drop_zero = FALSE, maxCombine = 0, peakPos = NULL)
}
\arguments{
\item{x}{An object of type \code{aligns}.}

\item{drop_zero}{Drop columns where all values are zero?}

\item{maxCombine}{How many adjacent columns to consider for merging.}

\item{peakPos}{Integer vector of column indices in the \code{cs} grid to snap
peaks to. Used to align new spectra to the same features as a reference
matrix (see 'Details').}
}
\value{
A numeric matrix with chemical shifts as rownames and spectrum names as
colnames (the transpose of \code{\link[=si_mat]{si_mat()}}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated in favour of \code{\link[=si_mat]{si_mat()}}, which returns the same data with
spectra in rows and features (chemical shifts) in columns.
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
