% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R
\name{install_mdrb}
\alias{install_mdrb}
\title{Install Rust Backend}
\usage{
install_mdrb()
}
\value{
NULL. Called for side effect of installing the Rust backend.
}
\description{
Installs the Rust backend and its dependencies.
}
