% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R
\name{install_mdrb}
\alias{install_mdrb}
\title{Install Rust Backend}
\usage{
install_mdrb(
  ask = TRUE,
  args_remotes = list(),
  args_mdrb = list(),
  verbose = TRUE
)
}
\arguments{
\item{ask}{Whether to ask for confirmation before attempting installation. Default is
TRUE.}

\item{args_remotes}{Additional arguments to pass to \code{\link[utils:install.packages]{utils::install.packages()}} when attempting
installation of \href{https://remotes.r-lib.org/index.html}{remotes}.}

\item{args_mdrb}{Additional arguments to pass to \code{\link[remotes:install_github]{remotes::install_github()}} when attempting
installation of mdrb.}

\item{verbose}{Whether to print messages to the console. Default is TRUE.}
}
\value{
NULL. Called for side effect of installing the Rust backend.
}
\description{
Installs metabodecon's Rust backend \href{https://github.com/spang-lab/mdrb}{mdrb}
via \code{\link[remotes:install_github]{remotes::install_github()}}. If remotes is not available, it is installed
first via \code{\link[utils:install.packages]{utils::install.packages()}}
}
\examples{
if (interactive()) try(install_mdrb())
}
\author{
Tobias Schmidt, 2024-2025: initial version.
}
