% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdrb.R
\name{install_mdrb}
\alias{install_mdrb}
\title{Install Rust Backend}
\usage{
install_mdrb(ask = TRUE, type = NULL, ...)
}
\arguments{
\item{ask}{Whether to ask for confirmation before attempting installation. Default is
TRUE.}

\item{type}{Installation type. Either "binary" or "source". Passed to
\code{\link[utils:install.packages]{utils::install.packages()}}. Leave at NULL to use the default type for your
operating system.}

\item{...}{Additional arguments to pass to \code{\link[utils:install.packages]{utils::install.packages()}} when attempting
installation of mdrb.}
}
\value{
NULL. Called for side effect of installing the Rust backend.
}
\description{
Installs metabodecon's Rust backend \href{https://github.com/spang-lab/mdrb}{mdrb}
from \href{https://spang-lab.r-universe.dev/mdrb}{R-Universe}.

lifecycle::badge("experimental")
}
\examples{
if (interactive()) try(install_mdrb())
}
\author{
2024-2025 Tobias Schmidt: initial version.
}
