% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{metabodecon-package}
\alias{metabodecon}
\alias{metabodecon-package}
\title{metabodecon: Deconvolution and Alignment of 1d NMR Spectra}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A framework for deconvolution, alignment and postprocessing of 1-dimensional (1d) nuclear magnetic resonance (NMR) spectra, resulting in a data matrix of aligned signal integrals. The deconvolution part uses the algorithm described in Koh et al. (2009) \doi{10.1016/j.jmr.2009.09.003}. The alignment part is based on functions from the 'speaq' package, described in Beirnaert et al. (2018) \doi{10.1371/journal.pcbi.1006018} and Vu et al. (2011) \doi{10.1186/1471-2105-12-405}. A detailed description and evaluation of an early version of the package, 'MetaboDecon1D v0.2.2', can be found in Haeckl et al. (2021) \doi{10.3390/metabo11070452}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/spang-lab/metabodecon/}
  \item \url{https://spang-lab.github.io/metabodecon/}
  \item Report bugs at \url{https://github.com/spang-lab/metabodecon/issues}
}

}
\author{
\strong{Maintainer}: Tobias Schmidt \email{tobias.schmidt331@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Martina Haeckl \email{martina.haeckl@stud.uni-regensburg.de} [copyright holder]
  \item Wolfram Gronwald \email{wolfram.gronwald@klinik.uni-regensburg.de} [copyright holder]
}

Other contributors:
\itemize{
  \item Yanren Linda Hu \email{yanren-linda.hu@klinik.uni-regensburg.de} [contributor]
}

}
\keyword{internal}
