% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/80_old_metabodecon1d.R
\name{MetaboDecon1D}
\alias{MetaboDecon1D}
\title{Deconvolute 1D NMR spectrum}
\usage{
MetaboDecon1D(
  filepath,
  filename = NA,
  file_format = "bruker",
  number_iterations = 10,
  range_water_signal_ppm = 0.1527692,
  signal_free_region = c(11.44494, -1.8828),
  smoothing_param = c(2, 5),
  delta = 6.4,
  scale_factor = c(1000, 1e+06),
  debug = FALSE
)
}
\arguments{
\item{filepath}{Complete path of the file folder (Notice for Bruker format: filepath need to be the spectrum folder containing one or more different spectra (e.g."C:/Users/Username/Desktop/spectra_from_bruker"))}

\item{filename}{Name of the NMR file. (Notice for Bruker format: filename need to be the name of your spectrum which is also the name of the folder) (Default: filename = NA to analyse more spectra at once)}

\item{file_format}{Format (bruker or jcampdx) of the NMR file. (Default: file_format = "bruker")}

\item{number_iterations}{Number of iterations for the approximation of the parameters for the Lorentz curves (Default: number_iterations=10)}

\item{range_water_signal_ppm}{Half width of the water artefact in ppm (Default: range_water_signal=0.1527692 (e.g. for urine NMR spectra))}

\item{signal_free_region}{Row vector with two entries consisting of the ppm positions for the left and right border of the signal free region of the spectrum. (Default: signal_free_region=c(11.44494, -1.8828))}

\item{smoothing_param}{Row vector with two entries consisting of the number of smoothing repeats for the whole spectrum and the number of data points (uneven) for the mean calculation (Default: smoothing_param=c(2,5))}

\item{delta}{Defines the threshold value to distinguish between signal and noise (Default: delta=6.4)}

\item{scale_factor}{Row vector with two entries consisting of the factor to scale the x-axis and the factor to scale the y-axis (Default: scale_factor=c(1000,1000000))}

\item{debug}{Logical value to activate the debug mode (Default: debug=FALSE)}
}
\value{
List containing
\itemize{
\item{\bold{filename}}
\item{\bold{x_values} (in datapoints)}
\item{\bold{x_values_ppm} (in ppm)}
\item{\bold{y_values} of the original spectrum}
\item{\bold{spectrum_superposition} y_values of the superposition of all generated Lorentz curves}
\item{\bold{mse_normed} is the mse value between the superposition of the Lorentz curves and the original spectrum which are standardized according to get a total area of 1}
\item{\bold{peak_triplets_left} (ppm x_values)}
\item{\bold{peak_triplets_middle} (ppm x_values)}
\item{\bold{peak_triplets_right} (ppm x_values)}
\item{\bold{index_peak_triplets_left} (index)}
\item{\bold{index_peak_triplets_middle} (index)}
\item{\bold{index_peak_triplets_right} (index)}
\item{\bold{integrals} for each generated Lorentz curve}
\item{\bold{signal_free_region} adjusted borders of signal free region}
\item{\bold{range_water_signal_ppm} adjusted range water signal in ppm}
\item{\bold{A} values of each Lorentz curve}
\item{\bold{lambda} values of each Lorentz curve}
\item{\bold{x_0} values of each Lorentz curve}
}

\bold{Notice:} The parameters A, lambda and x_0 to calculate the Lorentz curves are saved in parameters.txt and the approximated spectrum is saved in approximated_spectrum.txt under the file path.
}
\description{
Automatic deconvolution of a 1D NMR spectrum into several Lorentz curves and the integration of them. The NMR file need to be in Bruker format or jcamp-dx format.

Deprecated since metabodecon v1.2.0. Please use \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}} instead. See examples below for usage.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
The MetaboDecon1D package returns a list with i.a. the parameters A, lambda and x_0 to calculate the Lorentz curves.The Lorentz curves could be calculated by using the function calculate_lorentz_curves(). This returns a matrix containing the generated and approximated Lorentz curves for each real peak of the spectrum. Each row of the matrix depicts one Lorentz curve. The Lorentz curves could be visualised and saved by using the function plot_lorentz_curves_save_as_png(). The superposition of all Lorentz curves, which reconstructs the original spectrum, could also be visualised and saved with the plot_spectrum_superposition_save_as_png() function. For the analytical calculation of the Lorentz curves peak triplets for each peak are used. To visualise these peak triplets and to illustrate the impact of the threshold delta the function plot_triplets() is available. The integral values for each generated Lorentz curves are saved in a vector.

\bold{Notice}: It is feasible to load all spectra of a folder at once. Here the filename need to be "NA" which is the default value. One selected spectrum could then be used to adjust the parameters (signal_free_region and range_water_signal_ppm) for the analysis of all spectra. Furthermore it is possible to adjust these parameters for each spectrum separate.
}
\examples{
\donttest{
xds_path <- download_example_datasets()
urine <- file.path(xds_path, "bruker/urine")
urine_1 <- file.path(urine, "urine_1")

\dontrun{
# Deprecated since metabodecon v1.2.0. Please use generate_lorentz_curves()
# instead. Shown below.
urine_1_deconv <- MetaboDecon1D(urine, "urine_1")
urine_all_deconv <- MetaboDecon1D(urine)
}

urine_1_deconv <- generate_lorentz_curves(urine_1, ask = FALSE)
urine_all_deconv <- generate_lorentz_curves(urine, ask = FALSE)
}
}
\references{
Haeckl, M.; Tauber, P.; Schweda, F.; Zacharias, H.U.; Altenbuchinger, M.; Oefner, P.J.; Gronwald, W. An R-Package for the Deconvolution and Integration of 1D NMR Data: MetaboDecon1D. Metabolites 2021, 11, 452. https://doi.org/10.3390/metabo11070452
}
\seealso{
\code{\link{calculate_lorentz_curves}},
\code{\link{plot_triplets}},
\code{\link{plot_lorentz_curves_save_as_png}},
\code{\link{plot_spectrum_superposition_save_as_png}}
}
\author{
Martina Haeckl
}
