% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_util.R
\docType{data}
\name{metabodecon_classes}
\alias{metabodecon_classes}
\title{Metabodecon Classes}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
metabodecon_classes
}
\description{
Metabodecon introduces a set of classes to highlight the presence of certain elements in corresponding objects.

The order of elements may vary between different versions of Metabodecon, thus elements should always be accessed by name, for example, using \code{x$si} or \code{x[["cs"]]}.
A short description of each class is given in the listing below.
\itemize{
\item \code{spectrum}: One NMR spectrum
\item \code{decon0}: One deconvoluted NMR spectrum stored in \code{\link[=MetaboDecon1D]{MetaboDecon1D()}} format
\item \code{decon1}: One deconvoluted NMR spectrum stored in \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}} format
\item \code{decon2}: One deconvoluted NMR spectrum stored in \code{\link[=deconvolute_gspec]{deconvolute_gspec()}} format
\item \code{alignment}: One aligned NMR spectrum
\item \code{spectra}: List of multiple NMR spectra
\item \code{decons1}: List of multiple deconvoluted NMR stored in \code{\link[=MetaboDecon1D]{MetaboDecon1D()}} format
\item \code{decons2}: List of multiple deconvoluted NMR stored in \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}} format
\item \code{decons3}: List of multiple deconvoluted NMR stored in \code{\link[=deconvolute_gspec]{deconvolute_gspec()}} format
\item \code{alignments}: List of multiple aligned NMR spectra
}

More details can be found in Metabodecon's online documentation at \href{https://spang-lab.github.io/metabodecon/articles/Metabodecon-Classes.html}{Metabodecon Classes}.
}
\keyword{datasets}
