% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40_plot.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot Spectra}
\usage{
plot_spectra(
  x,
  mar = c(4.1, 4.1, 1.1, 0.1),
  peak_rng = get_ppm_range(x, show = FALSE)
)
}
\arguments{
\item{x}{An object of type decons0, decons1 or decons2. For details see
\href{https://spang-lab.github.io/metabodecon/articles/Metabodecon-Classes.html}{Metabodecon Classes}.}

\item{mar}{A numeric vector of length 4, which specifies the margins of the plot.}

\item{peak_rng}{A numeric vector of length 2, specifying the ppm range across
all peaks of the provided deconvoluted spectra.}
}
\value{
A plot of the deconvoluted spectra.
}
\description{
Plot a set of deconvoluted spectra.
}
\examples{
decons <- generate_lorentz_curves(
     data_path = sim,
     sfr = c(3.55, 3.35),
     wshw = 0,
     ask = FALSE,
     verbose = FALSE
)
plot_spectra(decons)

}
\seealso{
\code{\link[=plot_spectrum]{plot_spectrum()}} for a much more sofisticated plotting routine
suitable to plot a single spectrum.
}
