% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/51_plot_spectrum.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot Spectrum}
\usage{
plot_spectrum(
  decon,
  main = "",
  foc_rgn = c(0.4, 0.35),
  foc_unit = "fraction",
  foc_only = FALSE,
  sub_show = if (foc_only) FALSE else TRUE,
  sub_rgn = c(x1 = 0.05, x2 = 0.95, y1 = 0.2, y2 = 0.95),
  verbose = FALSE,
  rct_show = sub_show,
  rct_col = "black",
  rct_fill = transp("yellow"),
  xlab = "Chemical Shift [ppm]",
  ylab = "Signal Intensity [au]",
  mar = c(4, 4, if (main == "") 0 else 4, 1),
  line_col = "black",
  axis_col = "black",
  fill_col = NULL,
  box_col = "black",
  ysquash = if (sub_show) sub_rgn[3] * 0.96 else 0.96,
  trp_show = if (sub_show) FALSE else TRUE,
  lc_show = if (sub_show) FALSE else TRUE,
  sup_show = if (sub_show) FALSE else TRUE,
  sub_mar = c(2, 2, 0, 0),
  sub_line_col = line_col,
  sub_axis_col = axis_col,
  sub_fill_col = rct_fill,
  sub_box_col = rct_col,
  sub_trp_show = TRUE,
  sub_lc_show = TRUE,
  sub_sup_show = TRUE,
  sf_y_raw = 1e+06,
  sm_col = "blue",
  lc_col = "darkgrey",
  lc_lty = 1,
  lc_fill = transp(lc_col, 0.25),
  trp_col = rep(sm_col, 4),
  trp_pch = c(17, 4, 4, NA),
  sup_col = "red",
  sup_lty = 1,
  cnct_show = if (sub_show) TRUE else FALSE,
  cnct_col = rct_col
)
}
\arguments{
\item{decon}{An object as returned by \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}, containing the deconvolution data. Must include either \code{x_values_ppm} or \code{ppm} for the x-axis values, and either \code{y_values} or \code{y_smooth} for the y-axis values.}

\item{main}{Title of the plot.}

\item{foc_rgn}{Numeric vector specifying the start and end of focus region.}

\item{foc_unit}{Character string specifying the unit in which \code{foc_rgn} is given. Can be "fraction" or "ppm".}

\item{foc_only}{Logical. If TRUE, only the focused region is drawn. If FALSE, the full spectrum is drawn.}

\item{sub_show}{Logical. If TRUE, a sub figure is drawn within the main plot region. If FALSE, the sub figure is not drawn.}

\item{sub_rgn}{Either NULL or a numeric vector of the form \code{c(x1, x2, y1, y2)} giving the left/right/bottom/top coordinates of the sub figure region in "normalized plot coordinates" (as described in \code{\link[graphics:convertXY]{graphics::grconvertX()}}). If provided, the focused region is drawn as sub figure within the main plot region. Setting \code{sub_rgn} to NULL will prevent the sub figure from being drawn.}

\item{verbose}{Logical. If TRUE, print additional information.}

\item{rct_show}{Logical. If TRUE, the focus region is shown as a rectangle.}

\item{rct_col}{Border color of the rectangle around the focus region.}

\item{rct_fill}{Background color of the rectangle around the focus region.}

\item{xlab}{Label for the x-axis.}

\item{ylab}{Label for the y-axis.}

\item{mar}{Number of lines below/left/above/right plot region (used for axis annotations).}

\item{line_col}{Color of raw signal intensities.}

\item{axis_col}{Color of tickmarks and ticklabels.}

\item{fill_col}{Background color of the plot region.}

\item{box_col}{Border color of the box surrounding the plot region.}

\item{ysquash}{Fraction of plot height to squash y-values into. Useful in combination with \code{sub_rgn} to prevent the spectrum lines from overlapping with the sub figure showing the focused region.}

\item{trp_show}{Logical. If TRUE, the peak triplets are shown.}

\item{lc_show}{Logical. If TRUE, the Lorentzian Curves are shown.}

\item{sup_show}{Logical. If TRUE, the superposition of the Lorentzian Curves is shown.}

\item{sub_mar}{Margins of the sub figure.}

\item{sub_line_col}{Color of the lines in the sub figure.}

\item{sub_axis_col}{Color of the axis in the sub figure.}

\item{sub_fill_col}{Background color of the sub figure.}

\item{sub_box_col}{Border color of the box surrounding the sub figure.}

\item{sub_trp_show}{Logical. If TRUE, the peak triplets are shown in the sub figure.}

\item{sub_lc_show}{Logical. If TRUE, the Lorentzian Curves are shown in the sub figure.}

\item{sub_sup_show}{Logical. If TRUE, the superposition of the Lorentzian Curves is shown in the sub figure.}

\item{sf_y_raw}{Numeric value to divide raw signal intensities by before drawing.}

\item{sm_col}{Color of smoothed signal intensities.}

\item{lc_col}{Color of the Lorentzian Curves.}

\item{lc_lty}{Line type of the Lorentzian Curves.}

\item{lc_fill}{Color of the rectangles shown at center of each lorentzian curve. The width of the rectangle equals the half width at half height.}

\item{trp_col}{Vector of length 4 giving the colors for the peak triplets. The first three colors specify the color used for each peak-center, left-border and right-border and the fourth color used for any non-peak data point.}

\item{trp_pch}{Vector of length 4 giving the plotting characters for the peak triplets.}

\item{sup_col}{Color of the superposition of the Lorentzian Curves.}

\item{sup_lty}{Line type of the superposition of the Lorentzian Curves.}

\item{cnct_show}{Logical. If TRUE, connecting lines are drawn between the main and sub figure.}

\item{cnct_col}{Color of the lines connecting the main and sub figure.}
}
\value{
NULL. Called for side effect of plotting the spectrum, as sketched below.

\if{html}{\out{<div class="sourceCode">}}\preformatted{0__0.1_____________0.5_____________0.9__1
|    _______________________________    |0.9
|   |Focus Region  _/\\_   _         |   |
|   |       _    _/    \\_/ \\__      |   |
|   |    __/ \\__/             \\_    |   |
|   |___/_______________________\\___|   |0.5
|                    _______________    |
| Full /\\           |Focus /\\       |   |
| Spectrum          |Region  \\/\\    |   |
|    /    \\     /\\  |  /\\/      \\   |   |
|___/______\\___/__\\_|_/__________\\__|___|0
}\if{html}{\out{</div>}}
}
\description{
Plot a spectrum based on the provided deconvolution data and zoom in on a specific region of interest in the spectrum.

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# Prepare a deconvoluted spectrum as input
sim_01 <- metabodecon_file("sim/sim_01")
spec <- read_spectrum(sim_01)
decon <- generate_lorentz_curves(
    sim_01,
    sfr = c(3.42, 3.58), ws = 0, delta = 0.1,
    ask = FALSE, verbose = FALSE
)
testfunc <- function(figs = 1:6) {
    n <- length(figs)
    nr <- ceiling(sqrt(n))
    nc <- if (nr^2 > n) nr - 1 else nr
    opar <- par(mfrow = c(nr, nc))
    on.exit(par(opar))

    # Plot the full (non-deconvoluted) spectrum
    if (1 \%in\% figs) plot_spectrum(spec)

    # Focus on a specific region, specified in ppm or as fraction
    if (2 \%in\% figs) plot_spectrum(decon, foc_rgn = c(3.49, 3.45), foc_unit = "ppm")
    if (3 \%in\% figs) plot_spectrum(decon, foc_rgn = c(0.40, 0.30))

    # Change margin, color and position of the focus region
    if (4 \%in\% figs) {
        plot_spectrum(decon,
            sub_mar = c(4, 4, 0, 0),
            rct_fill = rgb(0.9, 0.5, 0.9, alpha = 0.1),
            rct_col = "violet",
            sub_rgn = c(x1 = 0.1, x2 = 0.9, y1 = 0.4, y2 = 0.9)
        )
    }

    # Remove connecting lines and fill colors
    if (5 \%in\% figs) {
        plot_spectrum(decon,
            rct_fill = NULL,
            sub_fill_col = NULL,
            cnct_col = NULL
        )
    }

    # Hide xlab and ylab
    if (6 \%in\% figs) {
        plot_spectrum(decon,
            xlab = "",
            ylab = "",
            mar = c(2, 2, 0, 1)
        )
    }
}
testfunc(3:4)
testfunc(1:6)
}
