% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_plot.R
\name{plot_spectrum}
\alias{plot_spectrum}
\title{Plot Spectrum}
\usage{
plot_spectrum(decon, focus = c(3.45, 3.55))
}
\arguments{
\item{decon}{An object as returned by \code{\link[=generate_lorentz_curves]{generate_lorentz_curves()}}, containing the deconvolution data. Must include either \code{x_values_ppm} or \code{ppm} for the x-axis values, and either \code{y_values} or \code{y_smooth} for the y-axis values.}

\item{focus}{A numeric vector of length 2 specifying the region of interest to highlight on the plot. The region is defined by its start and end points on the x-axis (in ppm).}
}
\value{
A plot is generated as a side effect, highlighting the specified focus region on the spectrum.
}
\description{
Plot a spectrum based on the provided deconvolution data and highlight a specific region of interest in the spectrum.

lifecycle::experimental()
}
\examples{
sim_01 <- metabodecon_file("sim/sim_01")
decon <- generate_lorentz_curves(
     sim_01, sfr = c(3.42, 3.58), ws = 0,
     ask = FALSE, verbose = FALSE
)
plot_spectrum(decon, focus = c(3.504, 3.450))
}
