% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_print.R
\name{print.spectrum}
\alias{print.spectrum}
\title{Print a Spectrum Object}
\usage{
\method{print}{spectrum}(x, name = FALSE, ...)
}
\arguments{
\item{x}{A spectrum object as returned by \code{\link[=make_spectrum]{make_spectrum()}}.}
}
\description{
Prints the name, path, type, magnetic field strength, number of data points, chemical shifts and frequencies of a spectrum object.
}
\examples{
si <- c(1, 1, 3, 7, 8, 3, 8, 5, 2, 1)
cs_max <- 14.8
cs_width <- 20.0
fq_ref <- 600.25 * 1e6
fq_width <- 12005
spectrum <- read_spectrum()
print.spectrum(spectrum)
}
