% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdm.R
\name{print.summary.mdm}
\alias{print.summary.mdm}
\title{Print summary.mdm object}
\usage{
\method{print}{summary.mdm}(x, ...)
}
\arguments{
\item{x}{Object of class \code{summary.mdm}.}

\item{...}{Additional arguments, currently unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Prints formatted output for objects of class \code{summary.mdm}.
}
\examples{
s <- structure(list(model = "svm", normalisation = "none", ref_len = 0),
  class = "summary.mdm"
)
print(s)
}
