% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/53_print.R
\name{print_methods}
\alias{print_methods}
\alias{print.spectrum}
\alias{print.ispec}
\alias{print.idecon}
\alias{print.decon1}
\alias{print.decon2}
\alias{print.spectra}
\alias{print.ispecs}
\alias{print.idecons}
\alias{print.decons1}
\alias{print.decons2}
\title{S3 Methods for Printing Metabodecon Objects}
\usage{
\method{print}{spectrum}(x, name = FALSE, ...)

\method{print}{ispec}(x, name = FALSE, ...)

\method{print}{idecon}(x, name = FALSE, ...)

\method{print}{decon1}(x, name = FALSE, ...)

\method{print}{decon2}(x, name = FALSE, ...)

\method{print}{spectra}(x, ...)

\method{print}{ispecs}(x, ...)

\method{print}{idecons}(x, ...)

\method{print}{decons1}(x, ...)

\method{print}{decons2}(x, ...)
}
\arguments{
\item{x}{The object to print.}

\item{name}{Logical. If TRUE, the name of the object is printed before the object.}

\item{...}{Not used. Only accepted to comply with generic \code{\link[base:print]{base::print()}}.}
}
\description{
S3 Methods for printing metabodecon objects as described in the \href{https://spang-lab.github.io/metabodecon/articles/}{Metabodecon Classes}.
}
\examples{
si <- c(1, 1, 3, 7, 8, 3, 8, 5, 2, 1)
cs_max <- 14.8
cs_width <- 20.0
fq_ref <- 600.25 * 1e6
fq_width <- 12005
spectrum <- read_spectrum()
print(spectrum)
}
