% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/61_sim.R
\docType{data}
\name{sim}
\alias{sim}
\title{The Sim Dataset}
\format{
A \href{metabodecon_classes}{spectra} object consisting of 16 \link{spectrum}
objects, where each spectrum object contains 2048 datapoints ranging from
3.60 to 3.29 ppm
}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}
}
\usage{
sim
}
\description{
A simulated dataset generated from the \href{https://spang-lab.github.io/metabodecon/articles/Datasets.html#Blood}{Blood}
dataset.
}
\keyword{datasets}
