% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/60_data.R
\docType{data}
\name{sim}
\alias{sim}
\title{The Sim Dataset}
\format{
A \href{metabodecon_classes}{spectra} object consisting of 16
\href{metabodecon_classes}{spectrum} objects, where each spectrum contains
2048 datapoints ranging from 3.60 to 3.29 ppm
}
\usage{
sim
}
\description{
A simulated dataset generated from the \href{https://spang-lab.github.io/metabodecon/articles/Datasets.html#Blood}{Blood}
dataset.
}
\keyword{datasets}
