% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_data.R
\name{tmpdir}
\alias{tmpdir}
\title{Temporary Session Directory}
\usage{
tmpdir(subdir = NULL, create = FALSE)
}
\arguments{
\item{subdir}{Optional subdirectory within the temporary session directory.}

\item{create}{Whether to create the directory if it does not yet exist.}
}
\value{
Returns the path to the temporary session directory.
}
\description{
Returns the path to metabodecon's temporary session directory. This directory
equals subdirectory 'metabodecon' of R's temporary session directory
\code{\link[base:tempfile]{base::tempdir()}} plus additional path normalization.
}
\examples{
tmpdir()
tmpdir("simulate_spectra")
}
\seealso{
\code{\link[=datadir_temp]{datadir_temp()}}
\code{\link[=datadir_persistent]{datadir_persistent()}}
}
